% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{aggregate.epichains}
\alias{aggregate.epichains}
\title{Aggregate cases in epichains objects according to a grouping variable}
\usage{
\method{aggregate}{epichains}(x, grouping_var = c("time", "generation"), ...)
}
\arguments{
\item{x}{An \code{\link{epichains}} object.}

\item{grouping_var}{The variable to group and count over. Options include
"time" and "generation".}

\item{...}{Other arguments passed to aggregate.}
}
\value{
An \verb{<epichains_aggregate_df>} object, which is basically a
\verb{<data.frame>}. The object stores the \code{chain_type = chains_tree} and
\code{grouping_var} attributes.
}
\description{
Aggregate cases in epichains objects according to a grouping variable
}
\examples{
set.seed(123)
chains <- simulate_tree(
  nchains = 10, statistic = "size",
  offspring_dist = "pois", stat_max = 10, serials_dist = function(x) 3,
  lambda = 2
)
chains

# Aggregate cases per time
aggregate(chains, grouping_var = "time")

# Aggregate cases per generation
aggregate(chains, grouping_var = "generation")

# Aggregate cases per both time and generation
aggregate(chains, grouping_var = "both")
}
\author{
James M. Azam
}
