% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{aggregate.epichains_tree}
\alias{aggregate.epichains_tree}
\title{Aggregate cases in \verb{<epichains_tree>} objects by "generation" or "time", if
present}
\usage{
\method{aggregate}{epichains_tree}(x, by = c("time", "generation"), ...)
}
\arguments{
\item{x}{An \verb{<epichains_tree>} object.}

\item{by}{The variable to aggregate by. Options include
"time" and "generation".}

\item{...}{Other arguments passed to aggregate.}
}
\value{
A \verb{<data.frame>} object of cases by \code{by}.
}
\description{
This function provides a quick way to create a time series of cases over
generation or time (if serials_dist was specified) from simulated
\verb{<epichains_tree>} objects.
}
\examples{
set.seed(123)
chains <- simulate_tree(
  ntrees = 10,
  statistic = "size",
  offspring_dist = "pois",
  stat_max = 10,
  generation_time = function(n) rep(3, n),
  lambda = 2
)
chains

# Aggregate cases per time
cases_per_time <- aggregate(chains, by = "time")
head(cases_per_time)

# Aggregate cases per generation
cases_per_gen <- aggregate(chains, by = "generation")
head(cases_per_gen)
}
\author{
James M. Azam
}
