% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_chain_tree_attribute}
\alias{check_chain_tree_attribute}
\title{Title}
\usage{
check_chain_tree_attribute(x)
}
\arguments{
\item{x}{An \code{\link{epichains}} object}
}
\description{
Title
}
\keyword{internal}
