% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_gen_interval_valid}
\alias{check_gen_interval_valid}
\title{Check if the gen_interval argument is specified as a function}
\usage{
check_gen_interval_valid(gen_interval)
}
\arguments{
\item{gen_interval}{The generation interval function; the name of a
user-defined named or anonymous function with only one argument \code{n},
representing the number of generation intervals to sample.}
}
\description{
Check if the gen_interval argument is specified as a function
}
\keyword{internal}
