% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_generation_time_valid}
\alias{check_generation_time_valid}
\title{Check if the generation_time argument is specified as a function}
\usage{
check_generation_time_valid(generation_time)
}
\arguments{
\item{generation_time}{The generation time function; the name
of a user-defined named or anonymous function with only one argument \code{n},
representing the number of generation times to sample.}
}
\description{
Check if the generation_time argument is specified as a function
}
\keyword{internal}
