% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_nchains_valid}
\alias{check_nchains_valid}
\title{Check that nchains is greater than 0 and not infinite}
\usage{
check_nchains_valid(nchains)
}
\arguments{
\item{nchains}{Number of chains to simulate.}
}
\description{
Check that nchains is greater than 0 and not infinite
}
\keyword{internal}
