% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_ntrees_valid}
\alias{check_ntrees_valid}
\title{Check that \code{ntrees} is greater than 0 and not infinity}
\usage{
check_ntrees_valid(ntrees)
}
\arguments{
\item{ntrees}{Number of trees to simulate.}
}
\description{
Check that \code{ntrees} is greater than 0 and not infinity
}
\keyword{internal}
