% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_offspring_valid}
\alias{check_offspring_valid}
\title{Check if offspring argument is specified as a character string}
\usage{
check_offspring_valid(offspring_dist)
}
\arguments{
\item{offspring_dist}{Offspring distribution: a character string
corresponding to the R distribution function (e.g., "pois" for Poisson,
where \code{\link{rpois}} is the R function to generate Poisson random
numbers).}
}
\description{
Check if offspring argument is specified as a character string
}
\keyword{internal}
