% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.adjust_next_gen}
\alias{.adjust_next_gen}
\title{Adjust next generation vector to match susceptible population size}
\usage{
.adjust_next_gen(next_gen, susc_pop)
}
\arguments{
\item{next_gen}{numeric; vector of next generation offspring}

\item{susc_pop}{numeric; susceptible population size}
}
\value{
numeric; adjusted next generation offspring vector
}
\description{
Adjust next generation vector to match susceptible population size
}
\keyword{internal}
