% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{.check_offspring_func_valid}
\alias{.check_offspring_func_valid}
\title{Check if constructed random number generator for offspring exists}
\usage{
.check_offspring_func_valid(roffspring_name)
}
\arguments{
\item{roffspring_name}{Constructed random offspring sampler: a character
string corresponding to the R distribution function (e.g., "rpois" for
Poisson.}
}
\description{
Check if constructed random number generator for offspring exists
}
\keyword{internal}
