% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{.check_sim_args}
\alias{.check_sim_args}
\title{Check inputs to \code{simulate_chains()} and \code{simulate_summary()}}
\usage{
.check_sim_args(
  func_name = c("simulate_chains", "simulate_summary"),
  index_cases,
  statistic,
  offspring_dist,
  stat_max,
  pop,
  percent_immune,
  generation_time_specified,
  generation_time = NULL,
  t0 = NULL,
  tf_specified,
  tf = NULL
)
}
\arguments{
\item{index_cases}{Number of index cases to simulate transmission chains for.}

\item{statistic}{\verb{<String>}; Chain statistic to track as the stopping
criteria for each chain being simulated when \code{stat_max} is not \code{Inf}.
Can be one of:
\itemize{
\item "size": the total number of cases produced by a chain before it goes
extinct.
\item "length": the total number of ancestors produced by a chain before
it goes extinct.
}}

\item{offspring_dist}{Offspring distribution: a function like the ones
provided by R to generate random numbers from given distributions (e.g.,
\code{\link{rpois}} for Poisson). More specifically, the function needs to
accept at least one argument, \code{n}, which is the number of random
numbers to generate. It can accept further arguments, which will be passed
on to the random number generating functions. Examples that can be provided
here are \code{rpois} for Poisson distributed offspring, \code{rnbinom} for negative
binomial offspring, or custom functions.}

\item{stat_max}{A cut off for the chain statistic (size/length) being
computed. Results above \code{stat_max} are set to \code{stat_max}. Defaults to \code{Inf}.}

\item{pop}{\verb{<Integer>}; Population size. Used alongside \code{percent_immune}. to
define the susceptible population. Defaults to \code{Inf}.}

\item{percent_immune}{\verb{<numeric>}; Percent of the population immune to
infection at the start of the simulation. Used alongside \code{pop} to initialise
the susceptible population. Accepted values lie between 0 and 1.
Defaults to 0.}

\item{generation_time}{The generation time function; the name
of a user-defined named or anonymous function with only one argument \code{n},
representing the number of generation times to sample.}

\item{t0}{Start time (if generation time is given); either a single value
or a vector of same length as \code{index_cases} (number of initial cases) with
corresponding initial times. Defaults to 0, meaning all cases started at
time 0.}

\item{tf}{Cut-off for the infection times (if generation time is given).
Defaults to \code{Inf}.}

\item{sim_func}{\if{html}{\out{<character>}}; The simulation function to check}
}
\value{
NULL; called for side effects
}
\description{
Check inputs to \code{simulate_chains()} and \code{simulate_summary()}
}
\keyword{internal}
