% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{.check_statistic_args}
\alias{.check_statistic_args}
\title{Check that the \code{statistic} and \code{stat_max} arguments are valid}
\usage{
.check_statistic_args(statistic, stat_max)
}
\arguments{
\item{statistic}{The chain statistic to track as the
stopping criteria for each chain being simulated when \code{stat_max} is not
\code{Inf}; A \verb{<string>}. It can be one of:
\itemize{
\item "size": the total number of cases produced by a chain before it goes
extinct.
\item "length": the total number of generations reached by a chain before
it goes extinct.
}}

\item{stat_max}{The cut off for the chain statistic (size/length) being
computed; A number coercible to integer. Results above \code{stat_max} are set to
\code{stat_max}; An \verb{<integer>}. Defaults to \code{Inf}.}
}
\value{
NULL; called for side effects
}
\description{
The function treats these two arguments as related and checks
them in one place to remove repeated checks in several places in the
package.
}
\keyword{internal}
