% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{.complementary_logprob}
\alias{.complementary_logprob}
\title{Calculates the complementary log-probability}
\usage{
.complementary_logprob(x)
}
\arguments{
\item{x}{A numeric vector of log-probabilities. Must be negative.}
}
\value{
A numeric value of the complementary log-probability.
}
\description{
Given x and norm, this calculates log(1-sum(exp(x)))
}
\author{
Sebastian Funk
}
\keyword{internal}
