% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_likelihoods.R
\name{.gborel_size_ll}
\alias{.gborel_size_ll}
\title{Log-likelihood of the size of chains with gamma-Borel offspring distribution}
\usage{
.gborel_size_ll(x, size, prob, mu)
}
\arguments{
\item{x}{A numeric vector of chain sizes.}

\item{size}{The dispersion parameter (often called \code{k} in ecological
applications); A positive number.}

\item{prob}{Probability of success (in the parameterisation with
\code{prob}, see also \code{\link[stats]{NegBinomial}}); A number between
0 and 1.}

\item{mu}{Mean; A positive number.}
}
\value{
A numeric vector of log-likelihood values.
}
\description{
Log-likelihood of the size of chains with gamma-Borel offspring distribution
}
\author{
Sebastian Funk James M. Azam
}
\keyword{internal}
