% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.get_infectible_offspring}
\alias{.get_infectible_offspring}
\title{Sample the number of infectible offspring from all possible offspring}
\usage{
.get_infectible_offspring(new_offspring, susc_pop, pop)
}
\arguments{
\item{new_offspring}{A vector of the possible new offspring per chain
produced by \code{\link[=.sample_possible_offspring]{.sample_possible_offspring()}}}

\item{pop}{\verb{<Integer>}; Population size. Used alongside \code{percent_immune}. to
define the susceptible population. Defaults to \code{Inf}.}
}
\value{
A vector of the number of offspring that can be infected given the
current susceptible population size
}
\description{
Sample susceptible offspring to be infected from all possible offspring.
This function is used internally, and input checking is not
performed here, only in the context where it is used. Using it directly
is not recommended.
}
\keyword{internal}
