% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.get_stat_threshold}
\alias{.get_stat_threshold}
\title{Extract the \code{stat_threshold} attribute from an \verb{<epichains>} object.}
\usage{
.get_stat_threshold(obj)
}
\arguments{
\item{obj}{An \if{html}{\out{<epichains>}} or <epichains_summary> object.}
}
\description{
Extract the \code{stat_threshold} attribute from an \verb{<epichains>} object.
}
\keyword{internal}
