% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.get_statistic_func}
\alias{.get_statistic_func}
\title{Return a function for calculating chain statistics}
\usage{
.get_statistic_func(chain_statistic)
}
\value{
A function for calculating chain statistics.
}
\description{
Return a function for calculating chain statistics
}
\keyword{internal}
