% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.get_susceptible_offspring}
\alias{.get_susceptible_offspring}
\title{Sample the number of susceptible offspring from all possible offspring}
\usage{
.get_susceptible_offspring(new_offspring, susc_pop, pop)
}
\arguments{
\item{new_offspring}{A numeric vector of the possible new offspring per
chain produced by \code{\link[=.sample_possible_offspring]{.sample_possible_offspring()}}.}

\item{pop}{Population size; An \verb{<Integer>}. Used alongside \code{percent_immune}
to define the susceptible population. Defaults to \code{Inf}.}
}
\value{
A numeric vector of the number of offspring that can be infected
given the current susceptible population size.
}
\description{
Sample susceptible offspring to be infected from all possible offspring.
This function is used internally, and input checking is not
performed here, only in the context where it is used. Using it directly
is not recommended.
}
\keyword{internal}
