% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{.is_epichains_summary}
\alias{.is_epichains_summary}
\title{Test if x is an \code{epichains_summary} object}
\usage{
.is_epichains_summary(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
Logical; \code{TRUE} if the object is an \verb{<epichains_summary>} and
\code{FALSE} otherwise.
}
\description{
Test if x is an \code{epichains_summary} object
}
\author{
James M. Azam
}
\keyword{internal}
