% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_likelihoods.R
\name{.pois_length_ll}
\alias{.pois_length_ll}
\title{Log-likelihood of the length of chains with Poisson offspring distribution}
\usage{
.pois_length_ll(x, lambda)
}
\arguments{
\item{x}{A numeric vector of chain lengths.}

\item{lambda}{The rate of the Poisson distribution. A single numeric value.
Must be positive.}
}
\value{
A numeric vector of log-likelihood values.
}
\description{
Log-likelihood of the length of chains with Poisson offspring distribution
}
\author{
Sebastian Funk
}
\keyword{internal}
