% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_likelihoods.R
\name{.pois_size_ll}
\alias{.pois_size_ll}
\title{Log-likelihood of the size of chains with Poisson offspring distribution}
\usage{
.pois_size_ll(x, lambda)
}
\arguments{
\item{x}{A numeric vector of sizes}

\item{lambda}{The rate of the Poisson distribution; A single numeric value.}
}
\value{
A numeric vector of log-likelihood values.
}
\description{
Log-likelihood of the size of chains with Poisson offspring distribution
}
\author{
Sebastian Funk
}
\keyword{internal}
