% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.rbinom_size}
\alias{.rbinom_size}
\title{Samples size (the number of trials) of a binomial distribution}
\usage{
.rbinom_size(n, x, prob)
}
\arguments{
\item{n}{The number of samples to generate. Must be a positive integer.}

\item{x}{The number of successes. Must be a positive integer.}

\item{prob}{The probability of success. A numeric between 0 and 1.}
}
\value{
A numeric vector of the sampled sizes.
}
\description{
Samples the size parameter from the binomial distribution with fixed x
(number of successes) and p (success probability)
}
\author{
Sebastian Funk
}
\keyword{internal}
