% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.rgen_length}
\alias{.rgen_length}
\title{Samples chain lengths with given observation probabilities}
\usage{
.rgen_length(n, x, prob)
}
\arguments{
\item{n}{The number of samples to generate. Must be a positive integer.}

\item{x}{A numeric vector of the observed chain lengths.}

\item{prob}{The probability of observation. A numeric between 0 and 1.}
}
\value{
A numeric vector of the sampled chain lengths.
}
\description{
Samples the length of a transmission chain where each individual element is
observed with binomial probability with parameters n (number of successes)
and p (success probability)
}
\author{
Sebastian Funk
}
\keyword{internal}
