% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.sample_possible_offspring}
\alias{.sample_possible_offspring}
\title{Sample all possible offspring for the next generation}
\usage{
.sample_possible_offspring(
  offspring_func,
  offspring_func_pars,
  n_offspring,
  chains
)
}
\arguments{
\item{offspring_func}{A function to sample offspring.}

\item{offspring_func_pars}{A list of parameters for the offspring function.}

\item{n_offspring}{A numeric vector of the number of offspring per chain.}

\item{chains}{Numeric indices of chains/infectors being simulated}
}
\value{
A numeric vector of the number of offspring per chain.
}
\description{
Sample next generation of offspring using offspring distribution and
associated parameters. This function is used internally, and input
checking is not performed here, only in the context where it is used.
Using it directly is not recommended.
}
\keyword{internal}
