% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.update_chain_stat}
\alias{.update_chain_stat}
\title{Determine and update the chain statistic being tracked}
\usage{
.update_chain_stat(stat_type, stat_latest, n_offspring)
}
\arguments{
\item{stat_type}{Chain statistic (size/length) to update; A character string.
Must be one of 'size' or 'length'.}

\item{stat_latest}{The latest chain statistic numeric vector to be updated.}

\item{n_offspring}{A vector of offspring per chain. A numeric vector
(coercible to integer).}
}
\value{
A vector of chain statistics (size/length). A numeric vector
coercible to integer.
}
\description{
Determine and update the chain statistic being tracked
}
\keyword{internal}
