% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{.validate_epichains}
\alias{.validate_epichains}
\title{Validate an \verb{<epichains>} object}
\usage{
.validate_epichains(x)
}
\arguments{
\item{x}{An \verb{<epichains>} object.}
}
\value{
Invisibly returns the object if it is valid.
}
\description{
Validate an \verb{<epichains>} object
}
\author{
James M. Azam
}
\keyword{internal}
