% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{epichains_tree}
\alias{epichains_tree}
\title{Create an \verb{<epichains_tree>} object}
\usage{
epichains_tree(tree_df, ntrees, statistic, offspring_dist, stat_max, track_pop)
}
\arguments{
\item{tree_df}{a \verb{<data.frame>} containing at least columns for
"infectee_id", "infector_id", and "generation". Also has optional columns
for "time", and "chain_id".}

\item{ntrees}{Number of initial cases used to generate the outbreak; Integer}

\item{statistic}{\verb{<String>}; Chain statistic to track as the stopping
criteria for each chain being simulated when \code{stat_max} is not \code{Inf}.
Can be one of:
\itemize{
\item "size": the total number of cases produced by a chain before it goes
extinct.
\item "length": the total number of ancestors produced by a chain before
it goes extinct.
}}

\item{offspring_dist}{Offspring distribution: a \verb{<character>} string
corresponding to the R distribution function (e.g., "pois" for Poisson,
where \code{\link{rpois}} is the R function to generate Poisson random
numbers).}

\item{stat_max}{A cut off for the chain statistic (size/length) being
computed. Results above \code{stat_max} are set to \code{stat_max}. Defaults to \code{Inf}.}

\item{track_pop}{Was the susceptible population tracked? Logical}
}
\value{
An \verb{<epichains_tree>} object
}
\description{
\code{epichains_tree()} constructs an \verb{<epichains_tree>} object, which is
inherently an \verb{<data.frame>} object that stores some of the inputs
passed to the \code{simulate_tree()} and \code{simulate_tree_from_pop()} and the
simulated output. The stored attributes are useful for downstream
analyses and reproducibility. This function checks the validity of the
object created to ensure it has the right columns and column types.

An \verb{<epichains_tree>} object contains a \verb{<data.frame>} of the simulated
outbreak tree with ids for each infector and infectee, generation, and
optionally, time, the number of initial cases used for the simulation,
the statistic that was tracked, and whether the susceptible population was
tracked.
}
\author{
James M. Azam
}
