% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{format.epichains}
\alias{format.epichains}
\title{Format method for epichains class}
\usage{
\method{format}{epichains}(x, ...)
}
\arguments{
\item{x}{epichains object}

\item{...}{further arguments passed to or from other methods}
}
\value{
Invisibly returns an \code{\link{epichains}}. Called for printing side-effects.
}
\description{
Format method for epichains class
}
\author{
James M. Azam
}
