% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{format.epichains_summary}
\alias{format.epichains_summary}
\title{Format method for \verb{<epichains_summary>} class}
\usage{
\method{format}{epichains_summary}(x, ...)
}
\arguments{
\item{x}{An \verb{<epichains_summary>} object}

\item{...}{further arguments passed to or from other methods}
}
\value{
Invisibly returns an \verb{<epichains_summary>}. Called for printing
side-effects.
}
\description{
Format method for \verb{<epichains_summary>} class
}
\author{
James M. Azam
}
