% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{format.epichains_tree}
\alias{format.epichains_tree}
\title{Format method for \verb{<epichains_tree>} class}
\usage{
\method{format}{epichains_tree}(x, ...)
}
\arguments{
\item{x}{An \verb{<epichains_tree>} object}

\item{...}{ignored}
}
\value{
Invisibly returns an \verb{<epichains_tree>}.
Called for printing side-effects.
}
\description{
Format method for \verb{<epichains_tree>} class
}
\author{
James M. Azam
}
