% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_chain_statistic_func}
\alias{get_chain_statistic_func}
\title{Return a function for calculating chain statistics}
\usage{
get_chain_statistic_func(chain_statistic)
}
\arguments{
\item{chain_statistic}{String; Statistic to calculate. Can be one of:
\itemize{
\item "size": the total number of offspring.
\item "length": the total number of ancestors.
}}
}
\value{
a function for calculating chain statistics
}
\description{
Return a function for calculating chain statistics
}
\keyword{internal}
