% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{head.epichains}
\alias{head.epichains}
\title{\code{head} method for \code{\link{epichains}} class}
\usage{
\method{head}{epichains}(x, ...)
}
\arguments{
\item{x}{An \code{\link{epichains}} object}

\item{...}{further arguments passed to or from other methods}
}
\value{
object of class \code{data.frame}
}
\description{
\code{head} method for \code{\link{epichains}} class
}
\details{
This returns the top rows of an \code{epichains} object. Note that the object
is originally sorted by \code{sim_id} and \code{ancestor} and the first
unknown ancestors (NA) have been dropped from
printing method. To view the full output, use \verb{as.data.frame(<object_name>)}.
}
\author{
James M. Azam
}
