% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{head.epichains_tree}
\alias{head.epichains_tree}
\alias{tail.epichains_tree}
\title{\code{head} and \code{tail} method for \verb{<epichains_tree>} class}
\usage{
\method{head}{epichains_tree}(x, ...)

\method{tail}{epichains_tree}(x, ...)
}
\arguments{
\item{x}{An \verb{<epichains_tree>} object}

\item{...}{further arguments passed to or from other methods}
}
\value{
Object of class \code{data.frame}
}
\description{
\code{head} and \code{tail} method for \verb{<epichains_tree>} class
}
\details{
This returns the top rows of an \verb{<epichains_tree>} object. Note that
the object is originally sorted by \code{sim_id} and \code{ancestor} and the first
unknown ancestors (NA) have been dropped from
printing method.

To view the full output, use \verb{as.data.frame(<object_name>)}.
}
\author{
James M. Azam
}
