% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention.R
\name{intvn_reduce_mean}
\alias{intvn_reduce_mean}
\title{Reduce the mean of the offspring distribution}
\usage{
intvn_reduce_mean(intvn_mean_reduction, offspring_dist, pars_list)
}
\arguments{
\item{intvn_mean_reduction}{A number between 0
and 1 for scaling/reducing the mean of \code{offspring_dist}. Serves as
population-level intervention. \code{intvn_mean_reduction} = 0
implies no intervention impact and \code{intvn_mean_reduction} = 1 implies full
impact.}

\item{offspring_dist}{Offspring distribution: a character string
corresponding to the R distribution function (e.g., "pois" for Poisson,
where \code{\link{rpois}} is the R function to generate Poisson random
numbers).}

\item{pars_list}{Parameter(s) for poisson or negative binomial offspring
distribution.}
}
\value{
List of the offspring distribution parameter(s) with the mean
scaled by \code{1 - intvn_mean_reduction}.
}
\description{
\code{intvn_reduce_mean()} is a helper for the \code{simulate_*()} functions. It
reduces/scales the mean of the offspring distribution in order to
mimic the impact of a population-level intervention. Currently, it can only
handle the poisson and negative binomial distributions and errors when other
offspring distributions are specified alongside the \code{intvn_mean_reduction}
argument.
}
\details{
\code{intvn_reduce_mean()} scales the mean of the offspring distribution
by \eqn{1 - {\sf intvn\_mean\_reduction}} so that the new mean is given as:
\deqn{(1 - {\sf intvn\_mean\_reduction}) \times {\sf mean,}} This
scaling when applied to the poisson and negative binomial offspring
distributions corresponds to the population-level reduction of R0 as
described in Lloyd-Smith et al, (2005). \code{intvn_reduce_mean()} is therefore
only implemented for the aforementioned distributions and errors when other
offspring distributions are specified along with the \code{intvn_mean_reduction}
argument in the \code{simulate_*()} functions.
}
\references{
Lloyd-Smith, J., Schreiber, S., Kopp, P. et al. Superspreading
and the effect of individual variation on disease emergence. Nature 438,
355–359 (2005). \doi{10.1038/nature04153}
}
\author{
James M. Azam
}
\keyword{internal}
