% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention.R
\name{intvn_scale_r0}
\alias{intvn_scale_r0}
\title{Set up intervention for simulation}
\usage{
intvn_scale_r0(r0_reduction, offspring_dist, pars_list)
}
\arguments{
\item{r0_reduction}{The intervention impact. A scalar between 0 and 1.
Scales the mean of \code{offspring_dist}. \code{r0_reduction} = 0 implies
no intervention impact and \code{r0_reduction} = 1 implies full impact.}

\item{offspring_dist}{Offspring distribution: a character string
corresponding to the R distribution function (e.g., "pois" for Poisson,
where \code{\link{rpois}} is the R function to generate Poisson random
numbers).}

\item{pars_list}{Parameter(s) for poisson or negative binomial offspring
distribution.}
}
\value{
List of the offspring distribution parameter(s) with the mean
scaled by \code{1 - intvn_scale_r0}.
}
\description{
\code{intvn_scale_r0()} is a helper for the \code{simulation_*} functions. It
modifies the relevant arguments of the offspring distribution in order to
mimic the impact of an intervention. In particular, it scales the mean of
the offspring distribution. Currently, it can only handle the poisson and
negative binomial distributions and errors when other offspring
distributions are specified alongside \code{intvn_scale_r0}.
}
\details{
\code{intvn_scale_r0()} scales the mean of the offspring distribution
by \eqn{1 - r0\_reduction} so that the new mean is given as:
\deqn{(1 - r0\_reduction) \times R_0,} where \eqn{R_0} is the
mean of the poisson and negative binomial distribution.
}
\author{
James M. Azam
}
\keyword{internal}
