% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{is_chains_tree}
\alias{is_chains_tree}
\title{Check if an epichains object has the \code{chains_tree} attribute}
\usage{
is_chains_tree(x)
}
\arguments{
\item{x}{An \code{\link{epichains}} object}
}
\description{
Check if an epichains object has the \code{chains_tree} attribute
}
\author{
James M. Azam
}
