% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{is_chains_vec}
\alias{is_chains_vec}
\title{Check if an epichains object has the \code{chains_vec} attribute}
\usage{
is_chains_vec(x)
}
\arguments{
\item{x}{An \code{\link{epichains}} object}
}
\description{
Check if an epichains object has the \code{chains_vec} attribute
}
\author{
James M. Azam
}
