% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{is_epichains_aggregate_df}
\alias{is_epichains_aggregate_df}
\title{Reports whether x is an "epichains_aggregate_df" object}
\usage{
is_epichains_aggregate_df(x)
}
\arguments{
\item{x}{An \code{\link{epichains}} object}
}
\value{
logical, \code{TRUE} if the object is an \code{epichains_aggregate_df} and
\code{FALSE} otherwise
}
\description{
Reports whether x is an "epichains_aggregate_df" object
}
\author{
James M. Azam
}
