% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{is_epichains_tree}
\alias{is_epichains_tree}
\title{Test if x is an \code{epichains_tree} object}
\usage{
is_epichains_tree(x)
}
\arguments{
\item{x}{An R object}
}
\value{
logical, \code{TRUE} if the object is an \code{epichains_tree} and \code{FALSE}
otherwise
}
\description{
Test if x is an \code{epichains_tree} object
}
\author{
James M. Azam
}
