% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{new_epichains_summary}
\alias{new_epichains_summary}
\title{Construct a \verb{<epichains_summary>} object}
\usage{
new_epichains_summary(
  chains_summary = vector(),
  chains_run = integer(),
  statistic = character(),
  stat_max = double(),
  intvn_mean_reduction = double()
)
}
\arguments{
\item{chains_summary}{a \verb{<vector>} of chain sizes and lengths.}

\item{chains_run}{Number of chains/cases used to generate the outbreak;
Integer}

\item{statistic}{String; Statistic (size/length) to calculate. Used to
determine stopping criteria for simulations when \code{stat_max} is finite.
Can be one of:
\itemize{
\item "size": the total number of offspring.
\item "length": the total number of ancestors.
}}

\item{stat_max}{A cut off for the chain statistic (size/length) being
computed. Results above the specified value, are set to this value.
Defaults to \code{Inf}.}

\item{intvn_mean_reduction}{A number between 0
and 1 for scaling/reducing the mean of \code{offspring_dist}. Serves as
population-level intervention. \code{intvn_mean_reduction} = 0
implies no intervention impact and \code{intvn_mean_reduction} = 1 implies full
impact.}
}
\description{
\code{new_epichains_summary()} constructs an \verb{<epichains_summary>} object from a
supplied \verb{<vector>} of chain sizes or lengths. It also stores extra
attributes passed as individual arguments.

\code{new_epichains_summary()} is meant to be lazy and performant, by creating
the object without checking the arguments for correctness. It is not safe
to call \code{new_epichains_summary()} on its own as is called within
\code{epichains_summary()} after the arguments have been checked. To create a
new \verb{<epichains_summary>} object safely, use \code{epichains_summary()}.
}
\author{
James M. Azam
}
\keyword{internal}
