% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{new_epichains_tree}
\alias{new_epichains_tree}
\title{Construct an \verb{<epichains_tree>} object}
\usage{
new_epichains_tree(
  tree_df,
  index_cases,
  statistic,
  offspring_dist,
  stat_max,
  track_pop
)
}
\arguments{
\item{tree_df}{a \verb{<data.frame>} containing at least columns for
"infectee_id", "infector_id", and "generation". Also has optional columns
for "time", and "chain_id".}

\item{index_cases}{Number of index cases to simulate transmission chains for.}

\item{statistic}{\verb{<String>}; Chain statistic to track as the stopping
criteria for each chain being simulated when \code{stat_max} is not \code{Inf}.
Can be one of:
\itemize{
\item "size": the total number of cases produced by a chain before it goes
extinct.
\item "length": the total number of ancestors produced by a chain before
it goes extinct.
}}

\item{offspring_dist}{Offspring distribution: a \verb{<character>} string
corresponding to the R distribution function (e.g., "pois" for Poisson,
where \code{\link{rpois}} is the R function to generate Poisson random
numbers).}

\item{stat_max}{A cut off for the chain statistic (size/length) being
computed. Results above \code{stat_max} are set to \code{stat_max}. Defaults to \code{Inf}.}

\item{track_pop}{Was the susceptible population tracked? Logical}
}
\description{
\code{new_epichains_tree()} constructs an \verb{<epichains_tree>} object from a
supplied \verb{<data.frame>} and extra attributes passed as individual arguments.
It is meant to be lazy and performant, by creating the object without
checking the arguments for correctness. It is not safe to call
\code{new_epichains_tree()} on its own as is called within \code{epichains_tree()}
after the arguments have been checked. To create an \verb{<epichains_tree>}
object, use \code{epichains_tree()}.
}
\author{
James M. Azam
}
\keyword{internal}
