% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_likelihoods.R
\name{offspring_ll}
\alias{offspring_ll}
\title{Log-likelihood of the summary (size/length) of chains with generic offspring
distribution}
\usage{
offspring_ll(chains, offspring_dist, statistic, nsim_offspring = 100, ...)
}
\arguments{
\item{chains}{Vector of chain summaries (sizes/lengths)}

\item{offspring_dist}{Offspring distribution: a character string
corresponding to the R distribution function (e.g., "pois" for Poisson,
where \code{\link{rpois}} is the R function to generate Poisson random
numbers).}

\item{statistic}{String; Statistic to calculate. Can be one of:
\itemize{
\item "size": the total number of offspring.
\item "length": the total number of ancestors.
}}

\item{nsim_offspring}{Number of simulations of the offspring distribution
for approximating the distribution of the chain statistic summary
(size/length)}

\item{...}{any parameters to pass to \code{\link{simulate_summary}}}
}
\value{
log-likelihood values
else raw likelihoods
}
\description{
The log-likelihoods are calculated with a crude approximation using simulated
chains by linearly approximating any missing values in the empirical
cumulative distribution function (ecdf).
}
\examples{
set.seed(123)
chain_size_ll <- offspring_ll(
  chains = c(1, 5, 6, 8, 7, 8, 10),
  offspring_dist = "pois",
  statistic = "size",
  lambda = 0.82
)
}
\seealso{
\code{\link[=simulate_summary]{simulate_summary()}} for simulating a summary of the transmission
chains statistic (without the tree of infections)
}
\author{
Sebastian Funk
}
