% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{plot.epichains}
\alias{plot.epichains}
\title{Plot epichains tree objects}
\usage{
\method{plot}{epichains}(x, ...)
}
\arguments{
\item{x}{An \code{\link{epichains}} object with a chains_tree attribute}

\item{...}{Other arguments passed to plot}
}
\value{
A plot of cases over time and generation
}
\description{
Plot epichains tree objects
}
\author{
James M. Azam
}
