% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{print.epichains}
\alias{print.epichains}
\title{Print an \code{\link{epichains}} object}
\usage{
\method{print}{epichains}(x, ...)
}
\arguments{
\item{x}{An \code{\link{epichains}} object.}

\item{...}{Other parameters passed to \code{\link[=print]{print()}}.}
}
\value{
Invisibly returns an \code{\link{epichains}}. Called for side-effects.
}
\description{
Print an \code{\link{epichains}} object
}
