% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{print.epichains_summary}
\alias{print.epichains_summary}
\title{Print an \verb{<epichains_summary>} object}
\usage{
\method{print}{epichains_summary}(x, ...)
}
\arguments{
\item{x}{An \verb{<epichains_summary>} object.}

\item{...}{Not used.}
}
\value{
Invisibly returns an \verb{<epichains_summary>}. Called for
side-effects.
}
\description{
Print an \verb{<epichains_summary>} object
}
\author{
James M. Azam
}
