% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{print.epichains_tree}
\alias{print.epichains_tree}
\title{Print an \verb{<epichains_tree>} object}
\usage{
\method{print}{epichains_tree}(x, ...)
}
\arguments{
\item{x}{An \verb{<epichains_tree>} object.}

\item{...}{Other parameters passed to \code{print()}.}
}
\value{
Invisibly returns an \verb{<epichains_tree>}. Called for
side-effects.
}
\description{
Print an \verb{<epichains_tree>} object
}
\author{
James M. Azam
}
