% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borel.r
\name{rborel}
\alias{rborel}
\title{Generate random numbers from the Borel distribution}
\usage{
rborel(n, mu, stat_max = Inf)
}
\arguments{
\item{n}{Number of random variates to generate.}

\item{mu}{A non-negative number for the poisson mean.}

\item{stat_max}{A stopping criterion for individual chain simulations; a
positive number coercible to integer. When any chain's cumulative statistic
reaches \code{stat_max}, that chain ends. It also serves as a censoring limit
so that results above the specified value, are set to \code{Inf}. Defaults to
\code{Inf}.}
}
\value{
A numeric vector of random numbers.
}
\description{
Random numbers are generated by simulating from a Poisson branching process
}
\examples{
set.seed(32)
rborel(5, 1)
}
\author{
Sebastian Funk
}
