% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borel.r
\name{rgborel}
\alias{rgborel}
\title{Generate random numbers from a Gamma-Borel mixture distribution}
\usage{
rgborel(n, size, prob, mu, stat_max = Inf)
}
\arguments{
\item{n}{Number of random variates to generate.}

\item{size}{the dispersion parameter (often called \code{k} in ecological
applications)}

\item{prob}{probability of success (in the parameterisation with
\code{prob}, see also \code{\link[stats]{NegBinomial}})}

\item{mu}{mean parameter}

\item{stat_max}{A stopping criterion for individual chain simulations; a
positive number coercible to integer. When any chain's cumulative statistic
reaches \code{stat_max}, that chain ends. It also serves as a censoring limit
so that results above the specified value, are set to \code{Inf}. Defaults to
\code{Inf}.}
}
\value{
Numeric vector of random numbers
}
\description{
Generate random numbers from a Gamma-Borel mixture distribution
}
\author{
Sebastian Funk
}
