% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_scenarios}
\alias{simulate_scenarios}
\title{Calculate transmission chain statistics across a range of scenarios and
divide into intervals}
\usage{
simulate_scenarios(statistic, offspring_dist, R_seq, k_seq, breaks, ...)
}
\arguments{
\item{statistic}{The chain statistic to track as the
stopping criteria for each chain being simulated when \code{stat_threshold} is not
\code{Inf}; A \verb{<string>}. It can be one of:
\itemize{
\item "size": the total number of cases produced by a chain before it goes
extinct.
\item "length": the total number of generations reached by a chain before
it goes extinct.
}}

\item{offspring_dist}{Offspring distribution: a \code{character} string  with the
name of the random number generator function. For example those
provided by R to generate random numbers from given distributions (e.g.,
\code{\link[=rpois]{rpois()}} for Poisson). If the \code{character} string does not match
a function then the function will error (see \code{\link[=match.fun]{match.fun()}} for details).
Examples that can be provided here are \code{"rpois"} for Poisson distributed
offspring, \code{"rnbinom"} for negative binomial offspring.}

\item{R_seq}{A \code{numeric} vector of reproduction number values to simulate
the branching process for.}

\item{k_seq}{A \code{numeric} vector of dispersion values to simulate the
branching process for. Only applicable for \code{offspring_dist = "rnbinom"}.}

\item{breaks}{either a numeric vector of two or more unique cut points or a
    single number (greater than or equal to 2) giving the number of
    intervals into which \code{x} is to be cut.}

\item{...}{\link{dots} Extra arguments to be passed to \code{\link[=simulate_chain_stats]{simulate_chain_stats()}}.
If argument does not match one from \code{\link[=simulate_chain_stats]{simulate_chain_stats()}} the function
will error.}
}
\value{
A \verb{<data.frame>}.
}
\description{
Calculate transmission chain statistics across a range of scenarios and
divide into intervals
}
\details{
The \code{offspring_dist} argument is not equivalent to the \code{offspring_dist}
argument in \code{\link[=simulate_chains]{simulate_chains()}} and \code{\link[=simulate_chain_stats]{simulate_chain_stats()}}, in those
functions \code{offspring_dist} should be a \verb{<function>}, whereas in
\code{simulate_scenarios()} \code{offspring_dist} should be a vector of \code{character}
strings which match function names.
}
\examples{
simulate_scenarios(
  statistic = c("size", "length"),
  offspring_dist = c("rpois", "rnbinom"),
  R_seq = seq(0.1, 0.5, 0.1),
  k_seq = seq(0.1, 0.5, 0.1),
  breaks = c(0, 2, 5, 10, 20, 50, 100, Inf)
)

# when R > 1 pass a stopping criteria to simulate_chain_stats()
simulate_scenarios(
  statistic = c("size", "length"),
  offspring_dist = c("rpois", "rnbinom"),
  R_seq = seq(0.5, 2.0, 0.5),
  k_seq = seq(0.1, 0.5, 0.1),
  breaks = c(0, 2, 5, 10, 20, 50, 100, Inf),
  stat_threshold = 10
)
}
