% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{simulate_tree}
\alias{simulate_tree}
\title{Simulate a tree of infections with a serial and offspring distributions}
\usage{
simulate_tree(
  nchains,
  statistic = c("size", "length"),
  offspring_dist,
  stat_max = Inf,
  serials_dist,
  t0 = 0,
  tf = Inf,
  ...
)
}
\arguments{
\item{nchains}{Number of chains to simulate.}

\item{statistic}{String; Statistic to calculate. Can be one of:
\itemize{
\item "size": the total number of offspring.
\item "length": the total number of ancestors.
}}

\item{offspring_dist}{Offspring distribution: a character string
corresponding to the R distribution function (e.g., "pois" for Poisson,
where \code{\link{rpois}} is the R function to generate Poisson random
numbers).}

\item{stat_max}{A cut off for the chain statistic (size/length) being
computed. Results above the specified value, are set to this value.
Defaults to \code{Inf}.}

\item{serials_dist}{The serial interval distribution function; the name
of a user-defined named or anonymous function with only one argument \code{n},
representing the number of serial intervals to generate. See details.}

\item{t0}{Start time (if serial interval is given); either a single value
or a vector of same length as \code{nchains} (number of simulations) with
initial times. Defaults to 0.}

\item{tf}{End time (if serial interval is given).}

\item{...}{Parameters of the offspring distribution as required by R.}
}
\value{
An \verb{<epichains>} object, which is basically a \verb{<data.frame>} with
columns \code{chain_id} (chain ID), \code{sim_id} (a unique ID within each simulation
for each individual), \code{ancestor}
(the ID of the ancestor of each individual), \code{generation}, and
\code{time} (of infection)
}
\description{
Simulate a tree of infections with a serial and offspring distributions
}
\details{
\code{simulate_tree()} simulates a branching process of the form:
WIP
}
\section{The serial interval (\code{serials_dist})}{
\subsection{Assumptions/disambiguation}{

In epidemiology, the generation interval is the duration between successive
infectious events in a chain of transmission. Similarly, the serial
interval is the duration between observed symptom onset times between
successive cases in a transmission chain. The generation interval is
often hard to observe because exact times of infection are hard to
measure hence, the serial interval is often used instead . Here, we
use the serial interval to represent what would normally be called the
generation interval, that is, the time between successive cases.

See References below for some literature on the subject.
}

\subsection{Specifying \code{serials_dist}}{

\code{serials_dist} must be specified as a named or
\href{https://en.wikipedia.org/wiki/Anonymous_function#R}{anonymous/inline/unnamed function} # nolint
with one argument.

For example, assuming we want to specify the serial interval
generator as a random log-normally distributed variable with
\code{meanlog = 0.58} and \code{sdlog = 1.58}, we could define a named function,
let's call it "serial_interval", with only one argument representing the
number of serial intervals to sample:
\code{serial_interval <- function(n){rlnorm(n, 0.58, 1.38)}},
and assign the name of the function to \code{serials_dist} in
\code{simulate_tree()} like so
\code{simulate_tree(..., serials_dist = serial_interval)},
where \code{...} are the other arguments to \code{simulate_tree()}.

Alternatively, we could assign an anonymous function to \code{serials_dist}
in the \code{simulate_tree()} call like so
\code{simulate_tree(..., serials_dist = function(n){rlnorm(n, 0.58, 1.38)})}, #nolint
where \code{...} are the other arguments to \code{simulate_tree()}.
}
}

\examples{
set.seed(123)
chains <- simulate_tree(
  nchains = 10,
  statistic = "size",
  offspring_dist = "pois",
  stat_max = 10,
  serials_dist = function(x) 3,
  lambda = 2
)
}
\references{
Lehtinen S, Ashcroft P, Bonhoeffer S. On the relationship
between serial interval, infectiousness profile and generation time.
J R Soc Interface. 2021 Jan;18(174):20200756.
doi: 10.1098/rsif.2020.0756. Epub 2021 Jan 6.
PMID: 33402022; PMCID: PMC7879757.

Fine PE. The interval between successive cases of an
infectious disease. Am J Epidemiol. 2003 Dec 1;158(11):1039-47.
doi: 10.1093/aje/kwg251. PMID: 14630599.
}
\seealso{
\code{\link[=simulate_summary]{simulate_summary()}} for simulating transmission chains
statistic without the full tree information.

\code{\link[=simulate_tree_from_pop]{simulate_tree_from_pop()}} for simulating transmission chains
from an initial susceptible population with initial immunity,
returning the full tree information ("sim_id",
"ancestor", "generation", and "time").
}
\author{
James M. Azam, Sebastian Funk
}
