% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{summary.epichains_summary}
\alias{summary.epichains_summary}
\title{Summary method for \verb{<epichains_summary>} class}
\usage{
\method{summary}{epichains_summary}(object, ...)
}
\arguments{
\item{object}{An \verb{<epichains_summary>} object}

\item{...}{ignored}
}
\value{
List of summaries
}
\description{
Summary method for \verb{<epichains_summary>} class
}
\author{
James M. Azam
}
