% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{tail.epichains}
\alias{tail.epichains}
\title{\code{tail} method for \code{\link{epichains}} class}
\usage{
\method{tail}{epichains}(x, ...)
}
\arguments{
\item{x}{An \code{\link{epichains}} object}

\item{...}{further arguments passed to or from other methods}
}
\description{
\code{tail} method for \code{\link{epichains}} class
}
\details{
This returns the bottom part of an \code{epichains} object. Note that the object
is originally sorted by \code{sim_id} and \code{infector_id} and the first
unknown ancestors (NA) have been dropped from
printing method. To view the full output, use \verb{as.data.frame(<object_name>)}.
}
\author{
James M. Azam
}
