% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{validate_epichains}
\alias{validate_epichains}
\title{\code{epichains} class validator}
\usage{
validate_epichains(x)
}
\arguments{
\item{x}{An \code{epichains} object}
}
\value{
No return.
}
\description{
\code{epichains} class validator
}
\author{
James M. Azam
}
