% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{validate_epichains_summary}
\alias{validate_epichains_summary}
\title{Validate an \verb{<epichains_summary>} object}
\usage{
validate_epichains_summary(x)
}
\arguments{
\item{x}{An \verb{<epichains_summary>} object}
}
\value{
No return.
}
\description{
Validate an \verb{<epichains_summary>} object
}
\author{
James M. Azam
}
