% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{validate_epichains_tree}
\alias{validate_epichains_tree}
\title{Validate an \verb{<epichains_tree>} object}
\usage{
validate_epichains_tree(x)
}
\arguments{
\item{x}{An \verb{<epichains_tree>} object}
}
\value{
No return.
}
\description{
Validate an \verb{<epichains_tree>} object
}
\author{
James M. Azam
}
