% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{aggregate.epichains}
\alias{aggregate.epichains}
\title{Aggregate cases in epichains objects according to a grouping variable}
\usage{
\method{aggregate}{epichains}(x, grouping_var = c("time", "generation", "both"), ...)
}
\arguments{
\item{x}{An \code{\link{epichains}} object.}

\item{grouping_var}{The variable to group and count over. Options include
"time", "generation", and "both".}

\item{...}{Other arguments passed to aggregate.}
}
\value{
If grouping_var is either "time" or "generation", a data.frame
with cases aggregated over \code{grouping_var}; If
\code{grouping_var = "both"}, a list of data.frames, the first being for
cases over time, and the second being for cases over generations.
}
\description{
Aggregate cases in epichains objects according to a grouping variable
}
\examples{
set.seed(123)
chains <- simulate_tree(nchains = 10, serials_sampler = function(x) 3,
offspring_sampler = "pois", lambda = 2, chain_stat_max = 10)
chains

# Aggregate cases per time
aggregate(chains, grouping_var = "time")

# Aggregate cases per generation
aggregate(chains, grouping_var = "generation")

# Aggregate cases per both time and generation
aggregate(chains, grouping_var = "both")
}
\author{
James M. Azam
}
