% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{aggregate.epichains}
\alias{aggregate.epichains}
\title{Aggregate cases in \verb{<epichains>} objects by "time" or "generation"}
\usage{
\method{aggregate}{epichains}(x, grouping_var = c("time", "generation"), ...)
}
\arguments{
\item{x}{An \verb{<epichains>} object.}

\item{grouping_var}{The variable to group and count over. Options include
"time" and "generation".}

\item{...}{Other arguments passed to aggregate.}
}
\value{
An \verb{<epichains_aggregate_df>} object, which is basically a
\verb{<data.frame>}. The object stores the \code{chain_type = chains_tree} and
\code{grouping_var} attributes.
}
\description{
Aggregate cases in \verb{<epichains>} objects by "time" or "generation"
}
\examples{
set.seed(123)
chains <- simulate_tree(
  ntrees = 10,
  statistic = "size",
  offspring_dist = "pois",
  stat_max = 10,
  generation_time = function(n) rep(3, n),
  lambda = 2
)
chains

# Aggregate cases per time
cases_per_time <- aggregate(chains, grouping_var = "time")
head(cases_per_time)

# Aggregate cases per generation
cases_per_gen <- aggregate(chains, grouping_var = "generation")
head(cases_per_gen)
}
\author{
James M. Azam
}
