% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{construct_offspring_ll_name}
\alias{construct_offspring_ll_name}
\title{Construct name of analytical function for estimating loglikelihood of
offspring}
\usage{
construct_offspring_ll_name(offspring_sampler, chain_statistic)
}
\value{
an analytical offspring likelihood function
}
\description{
Construct name of analytical function for estimating loglikelihood of
offspring
}
\keyword{internal}
