% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borel.r
\name{dborel}
\alias{dborel}
\title{Density of the Borel distribution}
\usage{
dborel(x, mu, log = FALSE)
}
\arguments{
\item{x}{A numeric vector of quantiles.}

\item{mu}{A non-negative number for the poisson mean.}

\item{log}{Logical; if TRUE, probabilities p are given as log(p).}
}
\value{
A numeric vector of the borel probability density.
}
\description{
Density of the Borel distribution
}
\examples{
set.seed(32)
dborel(1:5, 1)
}
\author{
Sebastian Funk
}
