% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.adjust_next_gen}
\alias{.adjust_next_gen}
\title{Adjust new offspring if it exceeds the susceptible population size}
\usage{
.adjust_next_gen(next_gen, susc_pop)
}
\arguments{
\item{next_gen}{numeric; vector of next generation offspring}

\item{susc_pop}{numeric; susceptible population size}
}
\value{
numeric; adjusted next generation offspring vector
}
\description{
This function is used internally, and input checking is not
performed here, only in the context where it is used. Using it directly
is not recommended.
}
\keyword{internal}
