% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{.check_sim_args}
\alias{.check_sim_args}
\title{Check inputs to \code{simulate_chains()} and \code{simulate_summary()}}
\usage{
.check_sim_args(
  nsims,
  statistic,
  offspring_dist,
  stat_max,
  pop,
  percent_immune
)
}
\arguments{
\item{nsims}{Number of cases to simulate transmission chains for.}

\item{statistic}{The chain statistic to track as the
stopping criteria for each chain being simulated when \code{stat_max} is not
\code{Inf}; A \verb{<string>}. It can be one of:
\itemize{
\item "size": the total number of cases produced by a chain before it goes
extinct.
\item "length": the total number of ancestors produced by a chain before
it goes extinct.
}}

\item{offspring_dist}{Offspring distribution: a \verb{<function>} like the ones
provided by R to generate random numbers from given distributions (e.g.,
\code{\link{rpois}} for Poisson). More specifically, the function needs to
accept at least one argument, \code{n}, which is the number of random
numbers to generate. It can accept further arguments, which will be passed
on to the random number generating functions. Examples that can be provided
here are \code{rpois} for Poisson distributed offspring, \code{rnbinom} for negative
binomial offspring, or custom functions.}

\item{stat_max}{The cut off for the chain statistic (size/length) being
computed; A number coercible to integer. Results above \code{stat_max} are set to
\code{stat_max}; An \verb{<integer>}. Defaults to \code{Inf}.}

\item{pop}{Population size; An \verb{<Integer>}. Used alongside \code{percent_immune}
to define the susceptible population. Defaults to \code{Inf}.}

\item{percent_immune}{Percent of the population immune to
infection at the start of the simulation; A \verb{<numeric>} between 0 and 1.
Used alongside \code{pop} to initialise the susceptible population. Defaults to
0.}
}
\value{
NULL; called for side effects
}
\description{
Check inputs to \code{simulate_chains()} and \code{simulate_summary()}
}
\keyword{internal}
