% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{.epichains}
\alias{.epichains}
\title{Create an \verb{<epichains>} object}
\usage{
.epichains(
  tree_df,
  index_cases,
  offspring_dist,
  track_pop,
  statistic = c("size", "length"),
  stat_max = Inf
)
}
\arguments{
\item{tree_df}{a \verb{<data.frame>} containing at least columns for
"infectee_id", "infector_id", and "generation". Also has optional columns
for "time", and "chain_id".}

\item{index_cases}{Number of index cases to simulate transmission chains for.}

\item{offspring_dist}{Offspring distribution: a function like the ones
provided by R to generate random numbers from given distributions (e.g.,
\code{\link{rpois}} for Poisson). More specifically, the function needs to
accept at least one argument, \code{n}, which is the number of random
numbers to generate. It can accept further arguments, which will be passed
on to the random number generating functions. Examples that can be provided
here are \code{rpois} for Poisson distributed offspring, \code{rnbinom} for negative
binomial offspring, or custom functions.}

\item{track_pop}{Was the susceptible population tracked? Logical}

\item{statistic}{\verb{<String>}; Chain statistic to track as the stopping
criteria for each chain being simulated when \code{stat_max} is not \code{Inf}.
Can be one of:
\itemize{
\item "size": the total number of cases produced by a chain before it goes
extinct.
\item "length": the total number of ancestors produced by a chain before
it goes extinct.
}}

\item{stat_max}{A cut off for the chain statistic (size/length) being
computed. Results above \code{stat_max} are set to \code{stat_max}. Defaults to \code{Inf}.}
}
\value{
An \verb{<epichains>} object
}
\description{
\code{epichains()} constructs an \verb{<epichains>} object, which is
inherently an \verb{<data.frame>} object that stores some of the inputs
passed to the \code{simulate_tree()} and \code{simulate_tree_from_pop()} and the
simulated output. The stored attributes are useful for downstream
analyses and reproducibility. This function checks the validity of the
object created to ensure it has the right columns and column types.

An \verb{<epichains>} object contains a \verb{<data.frame>} of the simulated
outbreak tree with ids for each infector and infectee, generation, and
optionally, time, the number of initial cases used for the simulation,
the statistic that was tracked, and whether the susceptible population was
tracked.
}
\author{
James M. Azam
}
\keyword{internal}
