% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{.epichains_summary}
\alias{.epichains_summary}
\title{Create an \verb{<epichains_summary>} object}
\usage{
.epichains_summary(
  chains_summary,
  index_cases,
  offspring_dist,
  statistic = c("size", "length"),
  stat_max = Inf
)
}
\arguments{
\item{chains_summary}{A numeric \verb{<vector>} of chain sizes and lengths.}

\item{index_cases}{Number of index cases to simulate transmission chains for.}

\item{offspring_dist}{Offspring distribution: a \verb{<function>} like the ones
provided by R to generate random numbers from given distributions (e.g.,
\code{\link{rpois}} for Poisson). More specifically, the function needs to
accept at least one argument, \code{n}, which is the number of random
numbers to generate. It can accept further arguments, which will be passed
on to the random number generating functions. Examples that can be provided
here are \code{rpois} for Poisson distributed offspring, \code{rnbinom} for negative
binomial offspring, or custom functions.}

\item{statistic}{The chain statistic to track as the
stopping criteria for each chain being simulated when \code{stat_max} is not
\code{Inf}; A \verb{<string>}. It can be one of:
\itemize{
\item "size": the total number of cases produced by a chain before it goes
extinct.
\item "length": the total number of ancestors produced by a chain before
it goes extinct.
}}

\item{stat_max}{The cut off for the chain statistic (size/length) being
computed; A number coercible to integer. Results above \code{stat_max} are set to
\code{stat_max}; An \verb{<integer>}. Defaults to \code{Inf}.}
}
\value{
An \verb{<epichains_summary>} object.
}
\description{
\code{epichains_summary()} constructs an \verb{<epichains_summary>} object.

An \verb{<epichains_summary>} object is a \verb{<vector>} of the simulated
tree sizes or lengths. It also stores information on the number of initial
cases used for the simulation, and the statistic that was tracked,
the intervention level.
}
\author{
James M. Azam
}
\keyword{internal}
