% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_likelihoods.R
\name{.offspring_ll}
\alias{.offspring_ll}
\title{Log-likelihood of the summary (size/length) of chains with generic offspring
distribution}
\usage{
.offspring_ll(x, offspring_dist, statistic, nsim_offspring = 100, ...)
}
\arguments{
\item{x}{Vector of chain summaries (sizes/lengths)}

\item{offspring_dist}{Offspring distribution: a function like the ones
provided by R to generate random numbers from given distributions (e.g.,
\code{\link{rpois}} for Poisson). More specifically, the function needs to
accept at least one argument, \code{n}, which is the number of random
numbers to generate. It can accept further arguments, which will be passed
on to the random number generating functions. Examples that can be provided
here are \code{rpois} for Poisson distributed offspring, \code{rnbinom} for negative
binomial offspring, or custom functions.}

\item{statistic}{\verb{<String>}; Chain statistic to track as the stopping
criteria for each chain being simulated when \code{stat_max} is not \code{Inf}.
Can be one of:
\itemize{
\item "size": the total number of cases produced by a chain before it goes
extinct.
\item "length": the total number of ancestors produced by a chain before
it goes extinct.
}}

\item{nsim_offspring}{Number of simulations of the offspring distribution
for approximating the distribution of the chain statistic summary
(size/length)}

\item{...}{any parameters to pass to \code{\link{simulate_summary}}}
}
\value{
log-likelihood values
}
\description{
The log-likelihoods are calculated with a crude approximation using simulated
chain summaries by linearly approximating any missing values in the empirical
cumulative distribution function (ecdf).
}
\examples{
set.seed(123)
chain_size_ll <- offspring_ll(
  x = c(1, 5, 6, 8, 7, 8, 10),
  offspring_dist = rpois,
  statistic = "size",
  lambda = 0.82
)
}
\seealso{
\code{\link[=simulate_summary]{simulate_summary()}} for simulating a summary of the transmission
chains statistic (without the tree of infections)
}
\author{
Sebastian Funk
}
\keyword{internal}
