% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{epichains_summary}
\alias{epichains_summary}
\title{Create an \verb{<epichains_summary>} object}
\usage{
epichains_summary(chains_summary, ntrees, statistic, offspring_dist, stat_max)
}
\arguments{
\item{chains_summary}{a \verb{<vector>} of chain sizes and lengths.}

\item{ntrees}{Number of initial cases used to generate the outbreak; Integer}

\item{statistic}{String; Statistic (size/length) to calculate. Used to
determine stopping criteria for simulations when \code{stat_max} is finite.
Can be one of:
\itemize{
\item "size": the total number of offspring.
\item "length": the total number of infectors.
}}

\item{offspring_dist}{Offspring distribution: a character string
corresponding to the R distribution function (e.g., "pois" for Poisson,
where \code{\link{rpois}} is the R function to generate Poisson random
numbers).}

\item{stat_max}{A cut off for the chain statistic (size/length) being
computed. Results above the specified value, are set to this value.
Defaults to \code{Inf}.}
}
\value{
An \verb{<epichains_summary>} object
}
\description{
\code{epichains_summary()} constructs an \verb{<epichains_summary>} object.

An \verb{<epichains_summary>} object is a \verb{<vector>} of the simulated
tree sizes or lengths. It also stores information on the number of initial
cases used for the simulation, and the statistic that was tracked,
the intervention level.
}
\author{
James M. Azam
}
