% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{epichains_summary}
\alias{epichains_summary}
\title{Create an \verb{<epichains_summary>} object}
\usage{
epichains_summary(
  chains_summary,
  chains_run = integer(),
  statistic = character(),
  stat_max = integer()
)
}
\arguments{
\item{chains_summary}{a \verb{<vector>} of chain sizes and lengths.}

\item{chains_run}{Number of chains/cases used to generate the outbreak;
Integer}

\item{statistic}{String; Statistic (size/length) to calculate. Used to
determine stopping criteria for simulations when \code{stat_max} is finite.
Can be one of:
\itemize{
\item "size": the total number of offspring.
\item "length": the total number of ancestors.
}}

\item{stat_max}{A cut off for the chain statistic (size/length) being
computed. Results above the specified value, are set to this value.
Defaults to \code{Inf}.}
}
\value{
An \verb{<epichains_summary>} object
}
\description{
\code{epichains_summary()} constructs an \verb{<epichains_summary>} object.

An \verb{<epichains_summary>} object is a \verb{<vector>} of the simulated
chain sizes or lengths. It also stores information on the
number of cases/chains used for the simulation, and the statistic that was
tracked, the intervention level.
}
\author{
James M. Azam
}
