% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{epichains_tree}
\alias{epichains_tree}
\title{Create an \verb{<epichains_tree>} object}
\usage{
epichains_tree(
  tree_df,
  ntrees = integer(),
  statistic = character(),
  stat_max = integer(),
  track_pop = logical()
)
}
\arguments{
\item{tree_df}{a \verb{<data.frame>} containing at least columns for
"infectee_id", "infector_id", and "generation". Also has optional columns
for "time", and "chain_id".}

\item{ntrees}{Number of trees to simulate.}

\item{statistic}{String; Statistic (size/length) to calculate. Used to
determine stopping criteria for simulations when \code{stat_max} is finite.
Can be one of:
\itemize{
\item "size": the total number of offspring.
\item "length": the total number of infectors.
}}

\item{stat_max}{A cut off for the chain statistic (size/length) being
computed. Results above the specified value, are set to this value.
Defaults to \code{Inf}.}

\item{track_pop}{Was the susceptible population tracked? Logical}
}
\value{
An \verb{<epichains_tree>} object
}
\description{
\code{epichains_tree()} constructs an \verb{<epichains_tree>} object, which is
inherently an \verb{<data.frame>} object that stores some of the inputs
passed to the \code{simulate_tree()} and \code{simulate_tree_from_pop()} and the
simulated output. The stored attributes are useful for downstream
analyses and reproducibility. This function checks the validity of the
object created to ensure it has the right columns and column types.

An \verb{<epichains_tree>} object contains a \verb{<data.frame>} of the simulated
outbreak tree with ids for each infector and infectee, generation, and
optionally, time, the number of initial cases used for the simulation,
the statistic that was tracked, and whether the susceptible population was
tracked.
}
\author{
James M. Azam
}
