% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{epichains_tree}
\alias{epichains_tree}
\title{Create an \verb{<epichains_tree>} object}
\usage{
epichains_tree(
  tree_df = data.frame(),
  chains_run = integer(),
  statistic = character(),
  stat_max = double(),
  intvn_mean_reduction = double(),
  track_pop = logical()
)
}
\arguments{
\item{tree_df}{a \verb{<data.frame>} containing at least columns for "chain_id",
"ancestor", and "generation". Also has optional columns for "time", and
"chain_id".}

\item{chains_run}{Number of chains/cases used to generate the outbreak;
Integer}

\item{statistic}{String; Statistic (size/length) to calculate. Used to
determine stopping criteria for simulations when \code{stat_max} is finite.
Can be one of:
\itemize{
\item "size": the total number of offspring.
\item "length": the total number of ancestors.
}}

\item{stat_max}{A cut off for the chain statistic (size/length) being
computed. Results above the specified value, are set to this value.
Defaults to \code{Inf}.}

\item{intvn_mean_reduction}{A number between 0
and 1 for scaling/reducing the mean of \code{offspring_dist}. Serves as
population-level intervention. \code{intvn_mean_reduction} = 0
implies no intervention impact and \code{intvn_mean_reduction} = 1 implies full
impact.}

\item{track_pop}{Was the susceptible population tracked; Logical}
}
\value{
An \verb{<epichains_tree>} object
}
\description{
\code{epichains_tree()} constructs an \verb{<epichains_tree>} object, which is
inherently an \verb{<data.frame>} object that stores some of the inputs
passed to the \code{simulate_tree()} and \code{simulate_tree_from_pop()} and the
simulated output. The stored attributes are useful for scenario
analyses where the inputs are required for downstream analyses.

An \verb{<epichains_tree>} object contains a \verb{<data.frame>} of the simulated
outbreak with ids for each case/chain and the chain the produced, the
number of cases/chains used for the simulation, the statistic that was
tracked, the intervention level, and whether the susceptible population was
tracked.
}
\author{
James M. Azam
}
