% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{format.epichains}
\alias{format.epichains}
\title{Format method for \verb{<epichains>} class}
\usage{
\method{format}{epichains}(x, ...)
}
\arguments{
\item{x}{An \verb{<epichains>} object}

\item{...}{ignored}
}
\value{
Invisibly returns an \verb{<epichains>}.
Called for printing side-effects.
}
\description{
Format method for \verb{<epichains>} class
}
\author{
James M. Azam
}
