% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_offspring_func}
\alias{get_offspring_func}
\title{Get offspring sampling function that takes into account susceptible
depletion}
\usage{
get_offspring_func(
  offspring_dist,
  n,
  susc,
  pop,
  mean_offspring,
  disp_offspring = NULL
)
}
\arguments{
\item{offspring_dist}{Offspring distribution: a character string
corresponding to the R distribution function (e.g., "pois" for Poisson,
where \code{\link{rpois}} is the R function to generate Poisson random
numbers).}

\item{n}{Number of items to sample}

\item{susc}{Susceptible population size (calculated
inside \code{\link{simulate_tree_from_pop}}  as pop - initial_immune)}

\item{pop}{The susceptible population.}
}
\value{
An offspring sampling function
}
\description{
Get offspring sampling function that takes into account susceptible
depletion
}
\keyword{internal}
