% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{head.epichains}
\alias{head.epichains}
\title{\code{head} method for \code{\link{epichains}} class}
\usage{
\method{head}{epichains}(x, ...)
}
\arguments{
\item{x}{An \code{\link{epichains}} object}

\item{...}{further arguments passed to or from other methods}
}
\value{
object of class \code{data.frame}
}
\description{
\code{head} method for \code{\link{epichains}} class
}
\author{
James M. Azam
}
