% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{head.epichains_tree}
\alias{head.epichains_tree}
\alias{tail.epichains_tree}
\title{\code{head} and \code{tail} method for \verb{<epichains_tree>} class}
\usage{
\method{head}{epichains_tree}(x, ...)

\method{tail}{epichains_tree}(x, ...)
}
\arguments{
\item{x}{An \verb{<epichains_tree>} object}

\item{...}{further arguments passed to or from other methods}
}
\value{
Object of class \code{data.frame}
}
\description{
\code{head} and \code{tail} method for \verb{<epichains_tree>} class
}
\details{
This returns the top rows of an \verb{<epichains_tree>} object. Note that
the object is originally sorted by \code{sim_id} and \code{infector_id} and the first
unknown infectors (NA) have been dropped from
printing method.

To view the full output, use \verb{as.data.frame(<object_name>)}.
}
\examples{
set.seed(32)
chains_pois_offspring <- simulate_chains(
  index_cases = 10,
  statistic = "size",
  offspring_dist = rpois,
  stat_max = 10,
  generation_time = function(n) rep(3, n),
  lambda = 2
)
head(chains_pois_offspring)
set.seed(32)
chains_pois_offspring <- simulate_chains(
  index_cases = 10,
  statistic = "size",
  offspring_dist = rpois,
  stat_max = 10,
  generation_time = function(n) rep(3, n),
  lambda = 2
)
tail(chains_pois_offspring)
}
\author{
James M. Azam
}
