% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{is_epichains}
\alias{is_epichains}
\title{Checks whether the object is an \code{epichains}}
\usage{
is_epichains(x)
}
\arguments{
\item{x}{An R object}
}
\value{
logical, \code{TRUE} if the object is an \code{epichains} and \code{FALSE}
otherwise
}
\description{
Checks whether the object is an \code{epichains}
}
\keyword{internal}
