% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{is_epichains_aggregate_df}
\alias{is_epichains_aggregate_df}
\title{Check if an object is of class "epichains_aggregate_df"}
\usage{
is_epichains_aggregate_df(x)
}
\arguments{
\item{x}{An \code{\link{epichains}} object}
}
\description{
Check if an object is of class "epichains_aggregate_df"
}
\keyword{internal}
