% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood}
\alias{likelihood}
\title{Estimate the log-likelihood/likelihood for observed branching processes}
\usage{
likelihood(
  chains,
  statistic = c("size", "length"),
  offspring_dist,
  nsim_obs,
  obs_prob = 1,
  log = TRUE,
  stat_max = Inf,
  exclude = NULL,
  individual = FALSE,
  ...
)
}
\arguments{
\item{chains}{Vector of chain summaries (sizes/lengths)}

\item{statistic}{\verb{<String>}; Chain statistic to track as the stopping
criteria for each chain being simulated when \code{stat_max} is not \code{Inf}.
Can be one of:
\itemize{
\item "size": the total number of cases produced by a chain before it goes
extinct.
\item "length": the total number of ancestors produced by a chain before
it goes extinct.
}}

\item{offspring_dist}{Offspring distribution: a function like the ones
provided by R to generate random numbers from given distributions (e.g.,
\code{\link{rpois}} for Poisson). More specifically, the function needs to
accept at least one argument, \code{n}, which is the number of random
numbers to generate. It can accept further arguments, which will be passed
on to the random number generating functions. Examples that can be provided
here are \code{rpois} for Poisson distributed offspring, \code{rnbinom} for negative
binomial offspring, or custom functions.}

\item{nsim_obs}{Number of simulations if the log-likelihood/likelihood is to
be approximated for imperfect observations.}

\item{obs_prob}{Observation probability (assumed constant)}

\item{log}{Logical; Should the log-likelihoods be transformed to
likelihoods? (Defaults to TRUE).}

\item{stat_max}{A cut off for the chain statistic (size/length) being
computed. Results above the specified value, are set to \code{Inf}.}

\item{exclude}{A vector of indices of the sizes/lengths to exclude from the
log-likelihood calculation.}

\item{individual}{If TRUE, a vector of individual log-likelihood/likelihood
contributions will be returned rather than the sum/product.}

\item{...}{any parameters to pass to \code{\link{simulate_summary}}}
}
\value{
If \code{log = TRUE}
\itemize{
\item A joint log-likelihood (sum of individual log-likelihoods), if
\code{individual == FALSE} (default) and \code{obs_prob == 1} (default), or
\item A list of individual log-likelihoods, if \code{individual == TRUE} and
\code{obs_prob == 1} (default), or
\item A list of individual log-likelihoods (same length as \code{nsim_obs}), if
\code{individual == TRUE} and \code{0 <= obs_prob < 1}, or
\item A vector of joint log-likelihoods (same length as \code{nsim_obs}), if
individual == FALSE and \code{0 <= obs_prob < 1} (imperfect observation).
}

If \code{log = FALSE}, the same structure of outputs as above are returned,
except that likelihoods, instead of log-likelihoods, are calculated in all
cases. Moreover, the joint likelihoods are the product, instead of the sum,
of the individual likelihoods.
}
\description{
Estimate the log-likelihood/likelihood for observed branching processes
}
\examples{
# example of observed chain sizes
set.seed(121)
# randomly generate 20 chains of size 1 to 10
chain_sizes <- sample(1:10, 20, replace = TRUE)
likelihood(
  chains = chain_sizes, statistic = "size",
  offspring_dist = rpois, nsim_obs = 100, lambda = 0.5
)
}
\seealso{
offspring_ll(), pois_size_ll(), nbinom_size_ll(), gborel_size_ll(),
pois_length_ll(), geom_length_ll()
}
\author{
Sebastian Funk
}
