% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood}
\alias{likelihood}
\title{Estimate the log-likelihood/likelihood for observed branching processes}
\usage{
likelihood(
  chains,
  statistic = c("size", "length"),
  offspring_dist,
  nsim_obs,
  log = TRUE,
  obs_prob = 1,
  stat_max = Inf,
  exclude = NULL,
  individual = FALSE,
  ...
)
}
\arguments{
\item{chains}{Vector of chain summaries (sizes/lengths)}

\item{statistic}{String; Statistic (size/length) to calculate. Used to
determine stopping criteria for simulations when \code{stat_max} is finite.
Can be one of:
\itemize{
\item "size": the total number of offspring.
\item "length": the total number of ancestors.
}}

\item{offspring_dist}{Offspring distribution: a character string
corresponding to the R distribution function (e.g., "pois" for Poisson,
where \code{\link{rpois}} is the R function to generate Poisson random
numbers).}

\item{nsim_obs}{Number of simulations if the log-likelihood/likelihood is to
be approximated for imperfect observations.}

\item{log}{Logical; Should the log-likelihoods be transformed to
likelihoods? (Defaults to TRUE).}

\item{obs_prob}{Observation probability (assumed constant)}

\item{stat_max}{A cut off for the chain statistic (size/length) being
computed. Results above the specified value, are set to \code{Inf}.}

\item{exclude}{A vector of indices of the sizes/lengths to exclude from the
log-likelihood calculation.}

\item{individual}{If TRUE, a vector of individual log-likelihood/likelihood
contributions will be returned rather than the sum/product.}

\item{...}{any parameters to pass to \code{\link{simulate_summary}}}
}
\value{
If \code{log = TRUE}
\itemize{
\item A joint log-likelihood (sum of individual log-likelihoods), if
\code{individual == FALSE} (default) and \code{obs_prob == 1} (default), or
\item A list of individual log-likelihoods, if \code{individual == TRUE} and
\code{obs_prob == 1} (default), or
\item A list of individual log-likelihoods (same length as \code{nsim_obs}), if
\code{individual == TRUE} and \code{0 <= obs_prob < 1}, or
\item A vector of joint log-likelihoods (same length as \code{nsim_obs}), if
individual == FALSE and \code{0 <= obs_prob < 1} (imperfect observation).
}

If \code{log = FALSE}, the same structure of outputs as above are returned,
except that likelihoods, instead of log-likelihoods, are calculated in all
cases. Moreover, the joint likelihoods are the product, instead of the sum,
of the individual likelihoods.
}
\description{
Estimate the log-likelihood/likelihood for observed branching processes
}
\examples{
# example of observed chain sizes
set.seed(121)
# randomly generate 20 chains of size 1 to 10
chain_sizes <- sample(1:10, 20, replace = TRUE)
likelihood(
  chains = chain_sizes, statistic = "size",
  offspring_dist = "pois", nsim_obs = 100, lambda = 0.5
)
}
\seealso{
offspring_ll(), pois_size_ll(), nbinom_size_ll(), gborel_size_ll(),
pois_length_ll(), geom_length_ll()
}
\author{
Sebastian Funk
}
