% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood}
\alias{likelihood}
\title{Estimate the (log) likelihood for observed branching processes}
\usage{
likelihood(
  chains,
  statistic = c("size", "length"),
  offspring_dist,
  nsim_obs,
  log = TRUE,
  obs_prob = 1,
  stat_max = Inf,
  exclude = NULL,
  individual = FALSE,
  ...
)
}
\arguments{
\item{chains}{Vector of sizes/lengths of transmission chains.}

\item{statistic}{String; Statistic to calculate. Can be one of:
\itemize{
\item "size": the total number of offspring.
\item "length": the total number of ancestors.
}}

\item{offspring_dist}{Offspring distribution: a character string
corresponding to the R distribution function (e.g., "pois" for Poisson,
where \code{\link{rpois}} is the R function to generate Poisson random
numbers).}

\item{nsim_obs}{Number of simulations if the likelihood is to be
approximated for imperfect observations.}

\item{log}{Logical; Should the results be log-transformed? (Defaults
to TRUE).}

\item{obs_prob}{Observation probability (assumed constant)}

\item{stat_max}{A cut off for the chain statistic (size/length) being
computed. Results above the specified value, are set to \code{Inf}.}

\item{exclude}{A vector of indices of the sizes/lengths to exclude from the
likelihood calculation.}

\item{individual}{If TRUE, a vector of individual (log)likelihood
contributions will be returned rather than the sum.}

\item{...}{Parameters for the offspring distribution.}
}
\value{
\itemize{
\item A log-likelihood, if \code{log = TRUE} (the default)
\item A vector of log-likelihoods, if \code{log = TRUE} (the default) and
\code{obs_prob < 1}, or
\item A list of individual log-likelihood contributions, if
\code{log = TRUE} (the default) and \code{individual = TRUE}.
else raw likelihoods, or vector of likelihoods
}
}
\description{
Estimate the (log) likelihood for observed branching processes
}
\examples{
# example of observed chain sizes
chain_sizes <- c(1, 1, 4, 7)
likelihood(
  chains = chain_sizes, statistic = "size",
  offspring_dist = "pois", nsim_obs = 100, lambda = 0.5
)
}
\seealso{
offspring_ll, pois_size_ll, nbinom_size_ll, gborel_size_ll,
pois_length_ll, geom_length_ll.
}
\author{
Sebastian Funk
}
