% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{new_epichains_summary}
\alias{new_epichains_summary}
\title{Construct a \verb{<epichains_summary>} object}
\usage{
new_epichains_summary(
  chains_summary,
  ntrees,
  statistic,
  offspring_dist,
  stat_max
)
}
\arguments{
\item{chains_summary}{a \verb{<vector>} of chain sizes and lengths.}

\item{ntrees}{Number of initial cases used to generate the outbreak; Integer}

\item{statistic}{\verb{<String>}; Chain statistic to track as the stopping
criteria for each chain being simulated when \code{stat_max} is not \code{Inf}.
Can be one of:
\itemize{
\item "size": the total number of cases produced by a chain before it goes
extinct.
\item "length": the total number of ancestors produced by a chain before
it goes extinct.
}}

\item{offspring_dist}{Offspring distribution: a \verb{<character>} string
corresponding to the R distribution function (e.g., "pois" for Poisson,
where \code{\link{rpois}} is the R function to generate Poisson random
numbers).}

\item{stat_max}{A cut off for the chain statistic (size/length) being
computed. Results above \code{stat_max} are set to \code{stat_max}. Defaults to \code{Inf}.}
}
\description{
\code{new_epichains_summary()} constructs an \verb{<epichains_summary>} object from a
supplied \verb{<vector>} of chain sizes or lengths. It also stores extra
attributes passed as individual arguments.

\code{new_epichains_summary()} is meant to be lazy and performant, by creating
the object without checking the arguments for correctness. It is not safe
to call \code{new_epichains_summary()} on its own as is called within
\code{epichains_summary()} after the arguments have been checked. To create a
new \verb{<epichains_summary>} object safely, use \code{epichains_summary()}.
}
\author{
James M. Azam
}
\keyword{internal}
