% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{new_epichains_tree}
\alias{new_epichains_tree}
\title{Construct a \verb{<epichains_tree>} object}
\usage{
new_epichains_tree(
  tree_df = data.frame(),
  chains_run = integer(),
  statistic = character(),
  stat_max = double(),
  intvn_mean_reduction = double(),
  track_pop = logical()
)
}
\arguments{
\item{tree_df}{a \verb{<data.frame>} containing at least columns for "chain_id",
"ancestor", and "generation". Also has optional columns for "time", and
"chain_id".}

\item{chains_run}{Number of chains/cases used to generate the outbreak;
Integer}

\item{statistic}{String; Statistic (size/length) to calculate. Used to
determine stopping criteria for simulations when \code{stat_max} is finite.
Can be one of:
\itemize{
\item "size": the total number of offspring.
\item "length": the total number of ancestors.
}}

\item{stat_max}{A cut off for the chain statistic (size/length) being
computed. Results above the specified value, are set to this value.
Defaults to \code{Inf}.}

\item{intvn_mean_reduction}{A number between 0
and 1 for scaling/reducing the mean of \code{offspring_dist}. Serves as
population-level intervention. \code{intvn_mean_reduction} = 0
implies no intervention impact and \code{intvn_mean_reduction} = 1 implies full
impact.}

\item{track_pop}{Was the susceptible population tracked; Logical}
}
\description{
\code{new_epichains_tree()} constructs an \verb{<epichains_tree>} object from a
supplied \verb{<data.frame>} and extra attributes passed as individual arguments.
It is meant to be lazy and performant, by creating the object without
checking the arguments for correctness. It is not safe to call
\code{new_epichains_tree()} on its own as is called within \code{epichains_tree()}
after the arguments have been checked. To create an \verb{<epichains_tree>}
object, use \code{epichains_tree()}.
}
\author{
James M. Azam
}
\keyword{internal}
