% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{offspring_ll}
\alias{offspring_ll}
\title{Likelihood of the length of chains with generic offspring distribution}
\usage{
offspring_ll(
  chains,
  offspring_dist,
  statistic,
  nsim_offspring = 100,
  log = TRUE,
  ...
)
}
\arguments{
\item{chains}{Vector of sizes/lengths}

\item{offspring_dist}{Offspring distribution: a character string
corresponding to the R distribution function (e.g., "pois" for Poisson,
where \code{\link{rpois}} is the R function to generate Poisson random
numbers).}

\item{statistic}{String; Statistic to calculate. Can be one of:
\itemize{
\item "size": the total number of offspring.
\item "length": the total number of ancestors.
}}

\item{nsim_offspring}{Number of simulations of the offspring distribution
for approximating the statistic (size/length) distribution}

\item{log}{Logical; Should the results be log-transformed? (Defaults
to TRUE).}

\item{...}{any parameters to pass to \code{\link{simulate_tree}}}
}
\value{
If \code{log = TRUE} (the default), log-likelihood values,
else raw likelihoods
}
\description{
The likelihoods are calculated with a crude approximation using simulated
chains by linearly approximating any missing values in the empirical
cumulative distribution function (ecdf).
}
\author{
Sebastian Funk
}
\keyword{internal}
