% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{offspring_ll}
\alias{offspring_ll}
\title{Likelihood of the length of chains with generic offspring distribution}
\usage{
offspring_ll(
  chains_observed,
  offspring_sampler,
  chain_statistic,
  nsim_offspring = 100,
  log_trans = TRUE,
  ...
)
}
\arguments{
\item{chains_observed}{Vector of sizes/lengths}

\item{offspring_sampler}{Offspring distribution: a character string
corresponding to the R distribution function (e.g., "pois" for Poisson,
where \code{\link{rpois}} is the R function to generate Poisson random
numbers).}

\item{chain_statistic}{Statistic given as \code{chains_observed}
("size" or "length" of chains).}

\item{nsim_offspring}{Number of simulations of the offspring distribution
for approximating the chain_statistic (size/length) distribution}

\item{log_trans}{Logical; Should the results be log-transformed? (Defaults
to TRUE).}

\item{...}{any parameters to pass to \code{\link{simulate_tree}}}
}
\value{
If \code{log_trans = TRUE} (the default), log-likelihood values,
else raw likelihoods
}
\description{
The likelihoods are calculated with a crude approximation using simulated
chains by linearly approximating any missing values in the empirical
cumulative distribution function (ecdf).
}
\author{
Sebastian Funk
}
\keyword{Compute}
\keyword{Cumulative}
\keyword{Distribution}
\keyword{Function}
\keyword{chains}
\keyword{empirical}
\keyword{internal}
\keyword{of}
\keyword{simulated}
\keyword{the}
