% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{plot.epichains}
\alias{plot.epichains}
\title{Plot epichains tree objects}
\usage{
\method{plot}{epichains}(x, ...)
}
\arguments{
\item{x}{An \code{\link{epichains}} object with a chains_tree attribute.}

\item{...}{Other arguments passed to plot.}
}
\value{
A plot of cases over time, generation, or both, depending on what
was aggregated over. See \code{?epichains::aggregate}.
}
\description{
This method accepts epichains aggregated through the \code{aggregate} method,
which returns an object of class \code{epichains_aggregate_df} with an
\code{aggregated_over} attribute that tells \code{plot()} which variable to plot.
}
\author{
James M. Azam
}
