% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{print.epichains}
\alias{print.epichains}
\title{Print an \verb{<epichains>} object}
\usage{
\method{print}{epichains}(x, ...)
}
\arguments{
\item{x}{An \verb{<epichains>} object.}

\item{...}{Other parameters passed to \code{print()}.}
}
\value{
Invisibly returns an \verb{<epichains>}. Called for
side-effects.
}
\description{
Print an \verb{<epichains>} object
}
\author{
James M. Azam
}
