% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{print.epichains_summary}
\alias{print.epichains_summary}
\title{Print an \verb{<epichains_summary>} object}
\usage{
\method{print}{epichains_summary}(x, ...)
}
\arguments{
\item{x}{An \verb{<epichains_summary>} object.}

\item{...}{Not used.}
}
\value{
Invisibly returns an \verb{<epichains_summary>}. Called for
side-effects.
}
\description{
Prints a summary of the \verb{<epichains_summary>} object. In particular, it
prints the number of chains simulated, and the range of
the statistic, represented as the maximum (\code{max_stat}) and minimum
(\code{min_stat}). If the minimum or maximum is infinite, it is represented as
\verb{>= stat_max} where \code{stat_max} is the value of the censoring limit. See
\code{?epichains_summary()} for the definition of \code{stat_max}.
}
\author{
James M. Azam
}
