% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borel.r
\name{rborel}
\alias{rborel}
\title{Generate random numbers from the Borel distribution}
\usage{
rborel(n, mu, infinite = Inf)
}
\arguments{
\item{n}{Number of random variates to generate.}

\item{mu}{mu parameter (the poisson mean); non-negative.}

\item{infinite}{A threshold above which the random number is set to
\code{Inf} as a form of truncation. This is important because \code{rborel()} uses
\code{simulate_summary()} internally with \code{offspring_dist = rpois}, so
when \code{mu >= 1}, the simulation could grow out of proportion and chain
summaries (the random numbers generated here) need to be truncated using
\code{infinite}.}
}
\value{
Vector of random numbers
}
\description{
Random numbers are generated by simulating from a Poisson branching process
}
\examples{
set.seed(32)
rborel(5, 1)
}
\author{
Sebastian Funk
}
