% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borel.R
\name{rborel}
\alias{rborel}
\title{Generate random numbers from the Borel distribution}
\usage{
rborel(n, mu, infinite = Inf)
}
\arguments{
\item{n}{Number of random variates to generate.}

\item{mu}{mu parameter (the poisson mean); non-negative.}

\item{infinite}{Any number to treat as infinite; simulations will be
stopped if this number is reached}
}
\value{
Vector of random numbers
}
\description{
Random numbers are generated by simulating from a Poisson branching process
}
\examples{
set.seed(32)
rborel(5, 1)
}
\author{
Sebastian Funk
}
