% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borel.r
\name{rgborel}
\alias{rgborel}
\title{Generate random numbers from a Gamma-Borel mixture distribution}
\usage{
rgborel(n, size, prob, mu, infinite = Inf)
}
\arguments{
\item{n}{Number of random variates to generate.}

\item{size}{the dispersion parameter (often called \code{k} in ecological
applications)}

\item{prob}{probability of success (in the parameterisation with
\code{prob}, see also \code{\link[stats]{NegBinomial}})}

\item{mu}{mean parameter}

\item{infinite}{Any number to treat as infinite; simulations will be
stopped if this number is reached}
}
\value{
Vector of random numbers
}
\description{
Generate random numbers from a Gamma-Borel mixture distribution
}
\author{
Sebastian Funk
}
