% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{simulate_chains}
\alias{simulate_chains}
\title{Simulate transmission chains}
\usage{
simulate_chains(
  nsims,
  statistic = c("size", "length"),
  offspring_dist,
  ...,
  stat_max = Inf,
  pop = Inf,
  percent_immune = 0,
  generation_time = NULL,
  t0 = 0,
  tf = Inf
)
}
\arguments{
\item{nsims}{Number of cases to simulate transmission chains for.}

\item{statistic}{The chain statistic to track as the
stopping criteria for each chain being simulated when \code{stat_max} is not
\code{Inf}; A \verb{<string>}. It can be one of:
\itemize{
\item "size": the total number of cases produced by a chain before it goes
extinct.
\item "length": the total number of ancestors produced by a chain before
it goes extinct.
}}

\item{offspring_dist}{Offspring distribution: a \verb{<function>} like the ones
provided by R to generate random numbers from given distributions (e.g.,
\code{\link{rpois}} for Poisson). More specifically, the function needs to
accept at least one argument, \code{n}, which is the number of random
numbers to generate. It can accept further arguments, which will be passed
on to the random number generating functions. Examples that can be provided
here are \code{rpois} for Poisson distributed offspring, \code{rnbinom} for negative
binomial offspring, or custom functions.}

\item{...}{Parameters of the offspring distribution as required by R.}

\item{stat_max}{The cut off for the chain statistic (size/length) being
computed; A number coercible to integer. Results above \code{stat_max} are set to
\code{stat_max}; An \verb{<integer>}. Defaults to \code{Inf}.}

\item{pop}{Population size; An \verb{<Integer>}. Used alongside \code{percent_immune}
to define the susceptible population. Defaults to \code{Inf}.}

\item{percent_immune}{Percent of the population immune to
infection at the start of the simulation; A \verb{<numeric>} between 0 and 1.
Used alongside \code{pop} to initialise the susceptible population. Defaults to
0.}

\item{generation_time}{The generation time function; the name
of a user-defined named or anonymous function with only one argument \code{n},
representing the number of generation times to sample.}

\item{t0}{Start time (if generation time is given); either a single value
or a vector of same length as \code{nsims} (number of initial cases) with
corresponding initial times. Defaults to 0, meaning all cases started at
time 0.}

\item{tf}{A number for the cut-off for the infection times (if generation
time is given); Defaults to \code{Inf}.}
}
\value{
An \verb{<epichains>} object, which is basically a \verb{<data.frame>}
with columns:
\itemize{
\item \code{index_case_active} - an ID for the index cases still active (links
each infector and infectee to the index case that seeded the chain),
\item \code{infectee} - a unique ID for each infectee.
\item \code{infector} - an ID for the infector of each infectee.
\item \code{generation} - a discrete time point during which infection occurs, and
optionally,
\item \code{time} - the time of infection.
}
}
\description{
Generates transmission chains using a simple branching process model that
that simulates the supplied number of independent simulations, using an
offspring distribution for each simulation, and an optional generation time
distribution function.

The simulations are controlled by customisable stopping criteria, including
a threshold chain size or length, and a generation time cut off. The
function also optionally accepts population related inputs such as the
population size (defaults to Inf) and percentage of the population
initially immune (defaults to 0).
}
\section{Calculating chain sizes and lengths}{
The function simulates the chain size for individual \eqn{i} at time
\eqn{t}, \eqn{I_{t, i}}, as:
\deqn{I_{t, i} = \sum_{i}^{I_{t-1}}X_{t, i},}
and the chain length/duration for individual \eqn{i} at time \eqn{t},
\eqn{L_{t, i}}, as:
\deqn{L_{t, i} = {\sf min}(1, X_{t, i}), }
where \eqn{X_{t, i}} is the secondary cases generated by individual \eqn{i}
at time \eqn{t}, and \eqn{I_{0, i} = L_{0, i} = 1}.

The distribution of secondary cases, \eqn{X_{t, i}} is modelled by the
offspring distribution (\code{offspring_dist}).
}

\section{Specifying \code{generation_time}}{
The argument \code{generation_time} must be specified as a function with
one argument, \code{n}.

For example, assuming we want to specify the generation time as a random
log-normally distributed variable with \code{meanlog = 0.58} and \code{sdlog = 1.58},
we could define a named function, let's call it "generation_time_fn",
with only one argument representing the number of generation times to
sample: \code{generation_time_fn <- function(n){rlnorm(n, 0.58, 1.38)}},
and assign the name of the function to \code{generation_time} in
the simulation function, i.e.
\code{`simulate_*`(..., generation_time = generation_time_fn)},
where \code{...} are the other arguments to \verb{simulate_*()} and * is a placeholder
for the rest of simulation function's name.

Alternatively, we could assign an anonymous function to \code{generation_time}
in the \verb{simulate_*()} call, i.e.
\code{simulate_*(..., generation_time = function(n){rlnorm(n, 0.58, 1.38)})}
OR \code{simulate_*(..., generation_time = \(n){rlnorm(n, 0.58, 1.38)})},
where \code{...} are the other arguments to \verb{simulate_*()}.
}

\examples{
# Using a Poisson offspring distribution and simulating from an infinite
# population up to chain size 10.
set.seed(32)
chains_pois_offspring <- simulate_chains(
  nsims = 10,
  statistic = "size",
  offspring_dist = rpois,
  stat_max = 10,
  generation_time = function(n) rep(3, n),
  lambda = 2
)
chains_pois_offspring

# Using a Negative binomial offspring distribution and simulating from a
# finite population up to chain size 10.
set.seed(32)
chains_nbinom_offspring <- simulate_chains(
  nsims = 10,
  pop = 100,
  percent_immune = 0,
  statistic = "size",
  offspring_dist = rnbinom,
  stat_max = 10,
  generation_time = function(n) rep(3, n),
  mu = 2,
  size = 0.2
)
chains_nbinom_offspring
}
\references{
Jacob C. (2010). Branching processes: their role in epidemiology.
International journal of environmental research and public health, 7(3),
1186–1204. \doi{https://doi.org/10.3390/ijerph7031204}

Blumberg, S., and J. O. Lloyd-Smith. 2013. "Comparing Methods for
Estimating R0 from the Size Distribution of Subcritical Transmission
Chains." Epidemics 5 (3): 131–45.
\doi{https://doi.org/10.1016/j.epidem.2013.05.002}.

Farrington, C. P., M. N. Kanaan, and N. J. Gay. 2003.
"Branching Process Models for Surveillance of Infectious Diseases
Controlled by Mass Vaccination.” Biostatistics (Oxford, England)
4 (2): 279–95. \doi{https://doi.org/10.1093/biostatistics/4.2.279}.
}
\author{
James M. Azam, Sebastian Funk
}
