% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{simulate_summary}
\alias{simulate_summary}
\title{Simulate transmission chains sizes/lengths without infection tree}
\usage{
simulate_summary(
  nchains,
  statistic = c("size", "length"),
  offspring_dist,
  stat_max = Inf,
  ...
)
}
\arguments{
\item{nchains}{Number of chains to simulate.}

\item{statistic}{String; Statistic to calculate. Can be one of:
\itemize{
\item "size": the total number of offspring.
\item "length": the total number of ancestors.
}}

\item{offspring_dist}{Offspring distribution: a character string
corresponding to the R distribution function (e.g., "pois" for Poisson,
where \code{\link{rpois}} is the R function to generate Poisson random
numbers).}

\item{stat_max}{A cut off for the chain statistic (size/length) being
computed. Results above the specified value, are set to \code{Inf}.}

\item{...}{Parameters of the offspring distribution as required by R.}
}
\description{
Simulate transmission chains sizes/lengths without infection tree
}
\section{Calculating chain sizes and lengths}{
The function simulates the chain size for individual \eqn{i} at time
\eqn{t}, \eqn{I_{t, i}}, as:
\deqn{I_{t, i} = \sum_{i}^{I_{t-1}}X_{t, i},}
and the chain length/duration for individual \eqn{i} at time \eqn{t},
\eqn{L_{t, i}}, as:
\deqn{L_{t, i} = {\sf min}(1, X_{t, i}), }
where \eqn{X_{t, i}} is the secondary cases generated by individual \eqn{i}
at time \eqn{t}, and \eqn{I_{0, i} = L_{0, i} = 1}.

The distribution of secondary cases, \eqn{X_{t, i}} is modelled by the
offspring distribution (\code{offspring_dist}).
}

\section{The serial interval (\code{serials_dist})}{
\subsection{Assumptions/disambiguation}{

In epidemiology, the generation interval is the duration between successive
infectious events in a chain of transmission. Similarly, the serial
interval is the duration between observed symptom onset times between
successive cases in a transmission chain. The generation interval is
often hard to observe because exact times of infection are hard to
measure hence, the serial interval is often used instead . Here, we
use the serial interval to represent what would normally be called the
generation interval, that is, the time between successive cases.

See References below for some literature on the subject.
}

\subsection{Specifying \code{serials_dist}}{

\code{serials_dist} must be specified as a named or
\href{https://en.wikipedia.org/wiki/Anonymous_function#R}{anonymous/inline/unnamed function} #nolint
with one argument.

For example, assuming we want to specify the serial interval
distribution as a random log-normally distributed variable with
\code{meanlog = 0.58} and \code{sdlog = 1.58}, we could define a named function,
let's call it "serial_interval", with only one argument representing the
number of serial intervals to sample:
\code{serial_interval <- function(n){rlnorm(n, 0.58, 1.38)}},
and assign the name of the function to \code{serials_dist} in
the simulation function like so
\code{`simulate_*`(..., serials_dist = serial_interval)},
where \code{...} are the other arguments to \verb{simulate_*()} and * is a placeholder
for the rest of simulation function's name.

Alternatively, we could assign an anonymous function to \code{serials_dist}
in the \verb{simulate_*()} call like so
\code{simulate_*(..., serials_dist = function(n){rlnorm(n, 0.58, 1.38)})},
where \code{...} are the other arguments to \verb{simulate_*()}.
}
}

\examples{
simulate_summary(
  nchains = 10, statistic = "size", offspring_dist = "pois",
  stat_max = 10, lambda = 2
)
}
\seealso{
\code{\link[=simulate_tree]{simulate_tree()}} for simulating the transmission chains,
returning the full tree information ("sim_id", "chain_id",
"ancestor", "generation", and optionally, "time").

\itemize{
\item \code{\link[=simulate_tree]{simulate_tree()}} for simulating transmission trees from an
initial number of infections.
\item \code{\link[=simulate_tree_from_pop]{simulate_tree_from_pop()}} for simulating transmission trees from a
susceptible or partially immune population.
}
}
\author{
James M. Azam, Sebastian Funk
}
