% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_summary}
\alias{simulate_summary}
\title{Simulate a vector of transmission chains sizes/lengths}
\usage{
simulate_summary(
  index_cases,
  statistic = c("size", "length"),
  offspring_dist,
  stat_max = Inf,
  ...
)
}
\arguments{
\item{index_cases}{Number of index cases to simulate transmission chains for.}

\item{statistic}{\verb{<String>}; Chain statistic to track as the stopping
criteria for each chain being simulated when \code{stat_max} is not \code{Inf}.
Can be one of:
\itemize{
\item "size": the total number of cases produced by a chain before it goes
extinct.
\item "length": the total number of ancestors produced by a chain before
it goes extinct.
}}

\item{offspring_dist}{Offspring distribution: a \verb{<character>} string
corresponding to the R distribution function (e.g., "pois" for Poisson,
where \code{\link{rpois}} is the R function to generate Poisson random
numbers).}

\item{stat_max}{A cut off for the chain statistic (size/length) being
computed. Results above the specified value, are set to \code{Inf}.}

\item{...}{Parameters of the offspring distribution as required by R.}
}
\description{
Generates a vector of transmission chain sizes and lengths with a
value for each index case. It uses a simple branching process model that
accepts an index number of cases that seed the outbreak, a distribution
of offspring per case, and a chain size or length/duration (statistic)
to track. It optionally accepts other population related inputs
such as the population size (defaults to Inf) and percentage of the
population initially immune (defaults to 0).
}
\section{Calculating chain sizes and lengths}{
The function simulates the chain size for individual \eqn{i} at time
\eqn{t}, \eqn{I_{t, i}}, as:
\deqn{I_{t, i} = \sum_{i}^{I_{t-1}}X_{t, i},}
and the chain length/duration for individual \eqn{i} at time \eqn{t},
\eqn{L_{t, i}}, as:
\deqn{L_{t, i} = {\sf min}(1, X_{t, i}), }
where \eqn{X_{t, i}} is the secondary cases generated by individual \eqn{i}
at time \eqn{t}, and \eqn{I_{0, i} = L_{0, i} = 1}.

The distribution of secondary cases, \eqn{X_{t, i}} is modelled by the
offspring distribution (\code{offspring_dist}).
\subsection{Specifying \code{generation_time}}{

The argument \code{generation_time} must be specified as a function with
one argument, \code{n}.

For example, assuming we want to specify the generation time as a random
log-normally distributed variable with \code{meanlog = 0.58} and \code{sdlog = 1.58},
we could define a named function, let's call it "generation_time_fn",
with only one argument representing the number of generation times to
sample: \code{generation_time_fn <- function(n){rlnorm(n, 0.58, 1.38)}},
and assign the name of the function to \code{generation_time} in
the simulation function, i.e.
\code{`simulate_*`(..., generation_time = generation_time_fn)},
where \code{...} are the other arguments to \verb{simulate_*()} and * is a placeholder
for the rest of simulation function's name.

Alternatively, we could assign an anonymous function to \code{generation_time}
in the \verb{simulate_*()} call, i.e.
\code{simulate_*(..., generation_time = function(n){rlnorm(n, 0.58, 1.38)})}
OR \code{simulate_*(..., generation_time = \(n){rlnorm(n, 0.58, 1.38)})},
where \code{...} are the other arguments to \verb{simulate_*()}.
}
}

\examples{
simulate_summary(
  index_cases = 10,
  statistic = "size",
  offspring_dist = "pois",
  stat_max = 10,
  lambda = 2
)
}
\author{
James M. Azam, Sebastian Funk
}
