% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{simulate_summary}
\alias{simulate_summary}
\title{Simulate transmission chains sizes/lengths without infection tree}
\usage{
simulate_summary(
  nchains,
  statistic = c("size", "length"),
  offspring_dist,
  stat_max = Inf,
  ...
)
}
\arguments{
\item{nchains}{Number of chains to simulate.}

\item{statistic}{String; Statistic to calculate. Can be one of:
\itemize{
\item "size": the total number of offspring.
\item "length": the total number of ancestors.
}}

\item{offspring_dist}{Offspring distribution: a character string
corresponding to the R distribution function (e.g., "pois" for Poisson,
where \code{\link{rpois}} is the R function to generate Poisson random
numbers).}

\item{stat_max}{A cut off for the chain statistic (size/length) being
computed. Results above the specified value, are set to \code{Inf}.}

\item{...}{Parameters of the offspring distribution as required by R.}
}
\description{
Simulate a summary of the transmission chain statistic
}
\examples{
simulate_summary(
  nchains = 10, statistic = "size", offspring_dist = "pois",
  stat_max = 10, lambda = 2
)
}
\seealso{
\code{\link[=simulate_tree]{simulate_tree()}} for simulating the transmission chains,
returning the full tree information ("sim_id", "chain_id",
"ancestor", "generation", and optionally, "time").

\code{\link[=simulate_tree_from_pop]{simulate_tree_from_pop()}} for simulating transmission chains
from an initial susceptible population with initial immunity,
returning the full tree information ("sim_id",
"ancestor", "generation", and "time").
}
\author{
James M. Azam, Sebastian Funk
}
