% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_summary}
\alias{simulate_summary}
\title{Simulate transmission chains sizes/lengths}
\usage{
simulate_summary(
  ntrees,
  statistic = c("size", "length"),
  offspring_dist,
  stat_max = Inf,
  ...
)
}
\arguments{
\item{ntrees}{Number of trees to simulate.}

\item{statistic}{String; Statistic (size/length) to calculate. Used to
determine stopping criteria for simulations when \code{stat_max} is finite.
Can be one of:
\itemize{
\item "size": the total number of offspring.
\item "length": the total number of infectors.
}}

\item{offspring_dist}{Offspring distribution: a character string
corresponding to the R distribution function (e.g., "pois" for Poisson,
where \code{\link{rpois}} is the R function to generate Poisson random
numbers).}

\item{stat_max}{A cut off for the chain statistic (size/length) being
computed. Results above the specified value, are set to \code{Inf}.}

\item{...}{Parameters of the offspring distribution as required by R.}
}
\description{
Simulate transmission chains sizes/lengths
}
\section{Calculating chain sizes and lengths}{
The function simulates the chain size for individual \eqn{i} at time
\eqn{t}, \eqn{I_{t, i}}, as:
\deqn{I_{t, i} = \sum_{i}^{I_{t-1}}X_{t, i},}
and the chain length/duration for individual \eqn{i} at time \eqn{t},
\eqn{L_{t, i}}, as:
\deqn{L_{t, i} = {\sf min}(1, X_{t, i}), }
where \eqn{X_{t, i}} is the secondary cases generated by individual \eqn{i}
at time \eqn{t}, and \eqn{I_{0, i} = L_{0, i} = 1}.

The distribution of secondary cases, \eqn{X_{t, i}} is modelled by the
offspring distribution (\code{offspring_dist}).
\subsection{Specifying \code{generation_time}}{

\code{generation_time} must be specified as a function with one argument.

For example, assuming we want to specify the generation time
as a random log-normally distributed variable with
\code{meanlog = 0.58} and \code{sdlog = 1.58}, we could define a named function,
let's call it "generation_time_fn", with only one argument representing the
number of generation times to sample:
\code{generation_time_fn <- function(n){rlnorm(n, 0.58, 1.38)}},
and assign the name of the function to \code{generation_time} in
the simulation function, i.e.
\code{`simulate_*`(..., generation_time = generation_time_fn)},
where \code{...} are the other arguments to \verb{simulate_*()} and * is a placeholder
for the rest of simulation function's name.

Alternatively, we could assign an anonymous function to \code{generation_time}
in the \verb{simulate_*()} call, i.e.
\code{simulate_*(..., generation_time = function(n){rlnorm(n, 0.58, 1.38)})},
where \code{...} are the other arguments to \verb{simulate_*()}.
}
}

\examples{
simulate_summary(
  ntrees = 10,
  statistic = "size",
  offspring_dist = "pois",
  stat_max = 10,
  lambda = 2
)
}
\seealso{
\itemize{
\item \code{\link[=simulate_tree]{simulate_tree()}} for simulating transmission trees from an
initial number of infections.
\item \code{\link[=simulate_tree_from_pop]{simulate_tree_from_pop()}} for simulating transmission trees from a
susceptible or partially immune population.
}
}
\author{
James M. Azam, Sebastian Funk
}
