% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{simulate_tree_from_pop}
\alias{simulate_tree_from_pop}
\title{Simulate transmission trees from a susceptible or partially immune
population}
\usage{
simulate_tree_from_pop(
  pop,
  offspring_dist = c("pois", "nbinom"),
  generation_time,
  initial_immune = 0,
  t0 = 0,
  tf = Inf,
  ...
)
}
\arguments{
\item{pop}{The susceptible population size.}

\item{offspring_dist}{Offspring distribution: a character string
corresponding to the R distribution function (e.g., "pois" for Poisson,
where \code{\link{rpois}} is the R function to generate Poisson random
numbers). Only supports "pois" and "nbinom".}

\item{generation_time}{The generation time function; the name
of a user-defined named or anonymous function with only one argument \code{n},
representing the number of generation times to sample.}

\item{initial_immune}{The number of initial immunes in the population.
Must be less than \code{pop} - 1.}

\item{t0}{Start time; Defaults to 0.}

\item{tf}{End time; Defaults to \code{Inf}.}

\item{...}{Parameters of the offspring distribution as required by R.}
}
\value{
An \verb{<epichains>} object, which is basically a \verb{<data.frame>} with
columns \code{sim_id} (a unique ID within each simulation for each infectee
in the chain), \code{infector_id}, \code{generation}, and \code{time} (of infection).
}
\description{
Simulate transmission trees from a susceptible or partially immune
population
}
\section{Offspring distributions}{
Currently, \code{offspring_dist} only supports "pois" & "nbinom".
Internally, the respective truncated poisson and negative binomial
distributions are used to avoid the situation where there are more cases
than susceptibles at any point.

The poisson model has mean, lambda, parametrised as:
\deqn{{\sf lambda} = \dfrac{{\sf lambda} \times ({\sf pop} -
{\sf initial\_immune} - 1)}{{\sf pop}}}

The negative binomial model, has mean, mu, parametrised as:
\deqn{{\sf mu} = \dfrac{{\sf mu} \times ({\sf pop} -
{\sf initial\_immune} - 1)}{{\sf pop}},}
and dispersion, size, parametrised as:
\deqn{{\sf size} = \dfrac{{\sf mu}}{{\sf size} - 1}.}
This is why \code{size} must be greater than 1.
}

\section{Differences with \code{simulate_tree()}}{
\code{simulate_tree_from_pop()} has a couple of key differences from
\code{simulate_tree()}:
\itemize{
\item the maximal chain statistic is limited by \code{pop} instead of
\code{stat_max} (in \code{simulate_tree()}),
\item \code{offspring_dist} can only handle "pois" and "nbinom".
}
}

\examples{
# Simulate with poisson offspring
simulate_tree_from_pop(
  pop = 100,
  offspring_dist = "pois",
  lambda = 0.5,
  generation_time = function(n) rep(3, n)
)

# Simulate with negative binomial offspring
simulate_tree_from_pop(
pop = 100, offspring_dist = "nbinom",
mu = 0.5,
size = 1.1,
generation_time = function(n) rep(3, n)
)
}
\seealso{
\itemize{
\item \code{\link[=simulate_tree]{simulate_tree()}} for simulating transmission trees from an
initial number of infections.
\item \code{\link[=simulate_summary]{simulate_summary()}} for simulating transmission chains
statistics (sizes or lengths) without the infection tree.
}
}
\author{
Flavio Finger, James M. Azam, Sebastian Funk
}
