% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{simulate_tree_from_pop}
\alias{simulate_tree_from_pop}
\title{Simulate a tree of infections from an initial susceptible population
with initial immunity}
\usage{
simulate_tree_from_pop(
  pop,
  offspring_dist = c("pois", "nbinom"),
  offspring_mean,
  offspring_disp,
  serial_dist,
  initial_immune = 0,
  t0 = 0,
  tf = Inf
)
}
\arguments{
\item{pop}{The susceptible population.}

\item{offspring_dist}{Offspring distribution: a character string
corresponding to the R distribution function (e.g., "pois" for Poisson,
where \code{\link{rpois}} is the R function to generate Poisson random
numbers).}

\item{offspring_mean}{The average number of secondary cases for each case.
Same as R0.}

\item{offspring_disp}{The dispersion parameter of the number of
secondary cases. Ignored if \code{offspring == "pois"}. Must be > 1 to
avoid division by 0 when calculating the size. See details and
\code{?rnbinom} for details on the parameterisation in Ecology.}

\item{serial_dist}{The serial interval. A function that takes one
parameter (\code{n}), the number of serial intervals to randomly sample. Value
must be >= 0.}

\item{initial_immune}{The number of initial immunes in the population.}

\item{t0}{Start time; Defaults to 0.}

\item{tf}{End time; Defaults to \code{Inf}.}
}
\value{
a data frame with columns \code{time}, \code{id} (a unique ID for each
individual element of the chain), \code{ancestor} (the ID of the ancestor
of each element), and \code{generation}.
}
\description{
Simulate a tree of infections from an initial susceptible population
with initial immunity
}
\section{Offspring distributions}{
Currently only "pois" & "nbinom" are supported. Internally truncated
distributions are used to avoid infecting more people than susceptibles
available.

The poisson model is parametrised so that:

lamda = offspring_mean * pop - initial_immune / pop

The negative binomial model is parametrised as:

mu = offspring_mean * pop - initial immune / pop, and
size = mu / (offspring_disp - 1). This is why offspring_disp must be greater
than 1.

simulate_tree_from_pop() has a couple of key different from simulate_tree():
\itemize{
\item the maximal chain statistic is limited by \code{pop} instead of
\code{stat_max} (in \code{simulate_tree()}),
\item it can only handle implemented offspring distributions ("pois" and
"nbinom").
}
}

\examples{
# Simulate with poisson offspring
simulate_tree_from_pop(
  pop = 100, offspring_dist = "pois",
  offspring_mean = 0.5, serial_dist = function(x) 3
)

# Simulate with negative binomial offspring
simulate_tree_from_pop(
  pop = 100, offspring_dist = "nbinom",
  offspring_mean = 0.5, offspring_disp = 1.1, serial_dist = function(x) 3
)
}
\author{
Flavio Finger

James M. Azam
}
