% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{simulate_tree_from_pop}
\alias{simulate_tree_from_pop}
\title{Simulate transmission trees from a susceptible or partially immune
population}
\usage{
simulate_tree_from_pop(
  pop,
  offspring_dist = c("pois", "nbinom"),
  serials_dist,
  initial_immune = 0,
  t0 = 0,
  tf = Inf,
  ...
)
}
\arguments{
\item{pop}{The susceptible population size.}

\item{offspring_dist}{Offspring distribution: a character string
corresponding to the R distribution function (e.g., "pois" for Poisson,
where \code{\link{rpois}} is the R function to generate Poisson random
numbers). Only supports "pois" and "nbinom".}

\item{serials_dist}{The serial interval distribution function; the name of a
user-defined named or anonymous function with only one argument, usually
called \code{n}, that returns a numeric vector of \code{n} randomly sampled serial
intervals. See details.}

\item{initial_immune}{The number of initial immunes in the population.
Must be less than \code{pop} - 1.}

\item{t0}{Start time; Defaults to 0.}

\item{tf}{End time; Defaults to \code{Inf}.}

\item{...}{Parameters of the offspring distribution as required by R.}
}
\value{
An \verb{<epichains>} object, which is basically a \verb{<data.frame>} with
columns \code{sim_id} (a unique ID within each simulation for each infectee
in the chain), \code{infector_id}, \code{generation}, and \code{time} (of infection).
}
\description{
Simulate transmission trees from a susceptible or partially immune
population
}
\section{Offspring distributions}{
Currently, \code{offspring_dist} only supports "pois" & "nbinom".
Internally, the respective truncated poisson and negative binomial
distributions are used to avoid the situation where there are more cases
than susceptibles at any point.

The poisson model has mean, lambda, parametrised as:
\deqn{{\sf lambda} = \dfrac{{\sf lambda} \times ({\sf pop} -
{\sf initial\_immune} - 1)}{{\sf pop}}}

The negative binomial model, has mean, mu, parametrised as:
\deqn{{\sf mu} = \dfrac{{\sf mu} \times ({\sf pop} -
{\sf initial\_immune} - 1)}{{\sf pop}},}
and dispersion, size, parametrised as:
\deqn{{\sf size} = \dfrac{{\sf mu}}{{\sf size} - 1}.}
This is why \code{size} must be greater than 1.
}

\section{Differences with \code{simulate_tree()}}{
\code{simulate_tree_from_pop()} has a couple of key differences from
\code{simulate_tree()}:
\itemize{
\item the maximal chain statistic is limited by \code{pop} instead of
\code{stat_max} (in \code{simulate_tree()}),
\item \code{offspring_dist} can only handle "pois" and "nbinom".
}
}

\section{The serial interval (\code{serials_dist})}{
\subsection{Assumptions/disambiguation}{

In epidemiology, the generation interval is the duration between successive
infectious events in a chain of transmission. Similarly, the serial
interval is the duration between observed symptom onset times between
successive cases in a transmission chain. The generation interval is
often hard to observe because exact times of infection are hard to
measure hence, the serial interval is often used instead . Here, we
use the serial interval to represent what would normally be called the
generation interval, that is, the time between successive cases.

See References below for some literature on the subject.
}

\subsection{Specifying \code{serials_dist}}{

\code{serials_dist} must be specified as a named or
\href{https://en.wikipedia.org/wiki/Anonymous_function#R}{anonymous/inline/unnamed function}
with one argument.

For example, assuming we want to specify the serial interval
distribution as a random log-normally distributed variable with
\code{meanlog = 0.58} and \code{sdlog = 1.58}, we could define a named function,
let's call it "serial_interval", with only one argument representing the
number of serial intervals to sample:
\code{serial_interval <- function(n){rlnorm(n, 0.58, 1.38)}},
and assign the name of the function to \code{serials_dist} in
the simulation function, i.e.
\code{`simulate_*`(..., serials_dist = serial_interval)},
where \code{...} are the other arguments to \verb{simulate_*()} and * is a placeholder
for the rest of simulation function's name.

Alternatively, we could assign an anonymous function to \code{serials_dist}
in the \verb{simulate_*()} call, i.e.
\code{simulate_*(..., serials_dist = function(n){rlnorm(n, 0.58, 1.38)})},
where \code{...} are the other arguments to \verb{simulate_*()}.
}
}

\examples{
# Simulate with poisson offspring
simulate_tree_from_pop(
  pop = 100,
  offspring_dist = "pois",
  lambda = 0.5,
  serials_dist = function(n) rep(3, n)
)

# Simulate with negative binomial offspring
simulate_tree_from_pop(
pop = 100, offspring_dist = "nbinom",
mu = 0.5,
size = 1.1,
serials_dist = function(n) rep(3, n)
)
}
\seealso{
\itemize{
\item \code{\link[=simulate_tree]{simulate_tree()}} for simulating transmission trees from an
initial number of infections.
\item \code{\link[=simulate_summary]{simulate_summary()}} for simulating transmission chains
statistics (sizes or lengths) without the infection tree.
}
}
\author{
Flavio Finger, James M. Azam, Sebastian Funk
}
