% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{summary.epichains}
\alias{summary.epichains}
\title{Summary method for epichains class}
\usage{
\method{summary}{epichains}(object, ...)
}
\arguments{
\item{object}{An \code{\link{epichains}} object}

\item{...}{further arguments passed to or from other methods}
}
\value{
data frame of information
}
\description{
Summary method for epichains class
}
\author{
James M. Azam
}
