% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{summary.epichains_summary}
\alias{summary.epichains_summary}
\title{Summary method for \verb{<epichains_summary>} class}
\usage{
\method{summary}{epichains_summary}(object, ...)
}
\arguments{
\item{object}{An \verb{<epichains_summary>} object.}

\item{...}{Not used.}
}
\value{
A list of chain summaries. The list contains the following
elements:
\itemize{
\item \code{index_cases}: the number of index cases used to simulate the chains.
\item \code{max_stat}: the maximum chain statistic (size/length) achieved by the
chains.
\item \code{min_stat}: the minimum chain statistic (size/length) achieved by the
chains.
}
}
\description{
Summary method for \verb{<epichains_summary>} class
}
\author{
James M. Azam
}
