% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{summary.epichains_tree}
\alias{summary.epichains_tree}
\title{Summary method for \code{epichains_tree} class}
\usage{
\method{summary}{epichains_tree}(object, ...)
}
\arguments{
\item{object}{An \code{epichains_tree} object}

\item{...}{further arguments passed to or from other methods}
}
\value{
List of summaries
}
\description{
Summary method for \code{epichains_tree} class
}
\author{
James M. Azam
}
