% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{tail.epichains}
\alias{tail.epichains}
\title{\code{tail} method for \code{\link{epichains}} class}
\usage{
\method{tail}{epichains}(x, ...)
}
\arguments{
\item{x}{An \code{\link{epichains}} object}

\item{...}{further arguments passed to or from other methods}
}
\description{
\code{tail} method for \code{\link{epichains}} class
}
\author{
James M. Azam
}
