% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epichains.R
\name{validate_epichains}
\alias{validate_epichains}
\title{\code{epichains} class validator}
\usage{
validate_epichains(x)
}
\arguments{
\item{x}{An \code{epichains} object}
}
\value{
Checks if an object is of class \code{epichains} and if so
checks that it's in the right format as a "data.frame" or vector.
}
\description{
\code{epichains} class validator
}
\author{
James M. Azam
}
\keyword{internal}
